### 2020-10-09 #2734 @kthyng
Hi @matthewfeickert! Thanks for your submission. One quick edit that will have to happen at some point: the references in your paper have () around them because they have been put in as [@reference] instead of @reference. Sometimes the () are justified of course but I see some in your paper that should be inline instead of parenthetical.
