### 2020-10-09 #2734 @matthewfeickert
> > Sometimes the () are justified of course but I see some in your paper that should be inline instead of parenthetical.
>
> Done.
