### 2020-10-09 #2734 @matthewfeickert
FWIW, as I'm used to seeing citations stylized as just the reference number in the bibliography

![example_screenshot](https://user-images.githubusercontent.com/5142394/95603677-3933f300-0a1c-11eb-84d1-f0aacb7039c8.png)

so if anything looks strange with the current switch to all inline references `@reference` then please let me know and I'll be happy to fix them.
