### 2020-10-09 9b19746
docs: Add contributors to website (#1095)
* Add list of all contributors as page to the website
* Update CONTRIBUTING.md to instruct contributors to add their names
