+++ docs/contributors.rst
@@ -1,0 +1,19 @@
+Contributors
+============
+
+``pyhf`` is openly developed and benefits from the contributions and feedback
+from its users.
+The ``pyhf`` dev team would like to thank all contributors to the project for
+their support and help.
+Thank you!
+
+Contributors include:
+
+- Jessica Forde
+- Ruggero Turra
+- Tadej Novak
+- Frank Sauerburger
+- Lars Nielsen
+- Kanishk Kalra
+- Nikolai Hartmann
+- Alexander Held
