### 2020-10-09 75f2a87
fix: Avoid deprecation warnings due to invalid escape sequences (#1047)
* Fix deprecation warnings due to invalid escape sequences by using raw strings
