+++ docs/contributors.rst
@@ -17,3 +17,4 @@ Contributors include:
 - Kanishk Kalra
 - Nikolai Hartmann
 - Alexander Held
+- Karthikeyan Singaravelan
