### 2020-10-10 6ce4669
feat: Add API to download likelihood patchset archives (#1046)
* Add download command to Python contrib API
* Add download command to pyhf CLI API and add contrib group
* Add pyhf custom exception if an untrusted archive host is given
* Add tests of Python and CLI API versions of contrib download
   - Thanks to Anthony Sottile (@asottile) for advice: https://youtu.be/PXu3KCMT3l4
* Add requests as dependency of contrib extra
* Add Python contrib API to docs
