### 2020-10-12 e551cd1
Set lower bound on CodeMetaPy
