### 2020-10-12 2ae3020
Add instructions on updateing metadata files
