+++ docs/development.rst
@@ -85,3 +85,20 @@ Once the PR is tagged with the label, the GitHub Actions bot will post a comment
 with information on the actions it will take once the PR is merged. When the PR
 has been reviewed, approved, and merged, the Tag Creator workflow will automatically
 create a new release with ``bump2version`` and then deploy the release to PyPI.
+
+Context Files and Archive Metadata
+----------------------------------
+
+The ``.zenodo.json`` and ``codemeta.json`` files have the version number
+automatically updated through ``bump2version``, though their additional metadata
+should be checked periodically by the dev team (probably every release).
+The `codemeta.json` file can be generated automatically **from a PyPI install**
+of ``pyhf`` using ``codemetapy``
+
+.. code-block:: bash
+
+  codemetapy --no-extras pyhf > codemeta.json
+
+though the ``author`` metadata will still need to be checked and revised by hand.
+The ``.zenodo.json`` is currently generated by hand, so it is worth using
+``codemeta.json`` as a guide to edit it.
