### 2020-10-12 6730984
docs: Add CodeMeta JSON-LD Context File (#1091)
* Add codemeta.json file generated from codemetapy
   - Control release number in codemeta.json with bumpversion
* Add codemetapy to 'develop' extra
* Add test to verify requirements in codemeta.json
* Update project description metadata across setup.cfg, .zenodo.json, and codemeta.json
