+++ setup.cfg
@@ -1,7 +1,7 @@
 [metadata]
 name = pyhf
 version = 0.5.2
-description = (partial) pure Python HistFactory implementation
+description = pure-Python HistFactory implementation with tensors and autodiff
 long_description = file: README.rst
 long_description_content_type = text/x-rst
 url = https://github.com/scikit-hep/pyhf
