+++ src/pyhf/mixins.py
@@ -42,3 +42,10 @@ class _ChannelSummaryMixin(object):
         self.samples = sorted(list(set(self.samples)))
         self.parameters = sorted(list(set(self.parameters)))
         self.modifiers = sorted(list(set(self.modifiers)))
+
+        self.channel_slices = {}
+        begin = 0
+        for c in self.channels:
+            end = begin + self.channel_nbins[c]
+            self.channel_slices[c] = slice(begin, end)
+            begin = end
