+++ tests/test_tensor.py
@@ -40,7 +40,14 @@ def test_simple_tensor_ops(backend):
     assert tb.tolist(tb.log(tb.exp(tb.astensor([2, 3, 4])))) == [2, 3, 4]
     assert tb.tolist(tb.abs(tb.astensor([-1, -2]))) == [1, 2]
     assert tb.tolist(tb.erf(tb.astensor([-2.0, -1.0, 0.0, 1.0, 2.0]))) == pytest.approx(
-        [-0.99532227, -0.84270079, 0.0, 0.84270079, 0.99532227,], 1e-7,
+        [
+            -0.99532227,
+            -0.84270079,
+            0.0,
+            0.84270079,
+            0.99532227,
+        ],
+        1e-7,
     )
     assert tb.tolist(
         tb.erfinv(tb.erf(tb.astensor([-2.0, -1.0, 0.0, 1.0, 2.0])))
@@ -83,13 +90,16 @@ def test_complex_tensor_ops(backend):
         1,
         1,
     ]
-    assert tb.tolist(
-        tb.where(
-            tb.astensor([1, 0, 1], dtype="bool"),
-            tb.astensor([1, 1, 1]),
-            tb.astensor([2, 2, 2]),
+    assert (
+        tb.tolist(
+            tb.where(
+                tb.astensor([1, 0, 1], dtype="bool"),
+                tb.astensor([1, 1, 1]),
+                tb.astensor([2, 2, 2]),
+            )
         )
-    ) == [1, 2, 1]
+        == [1, 2, 1]
+    )
 
 
 def test_ones(backend):
@@ -112,30 +122,39 @@ def test_zeros(backend):
 
 def test_broadcasting(backend):
     tb = pyhf.tensorlib
-    assert list(
-        map(
-            tb.tolist,
-            tb.simple_broadcast(
-                tb.astensor([1, 1, 1]), tb.astensor([2]), tb.astensor([3, 3, 3])
-            ),
+    assert (
+        list(
+            map(
+                tb.tolist,
+                tb.simple_broadcast(
+                    tb.astensor([1, 1, 1]), tb.astensor([2]), tb.astensor([3, 3, 3])
+                ),
+            )
         )
-    ) == [[1, 1, 1], [2, 2, 2], [3, 3, 3]]
-    assert list(
-        map(
-            tb.tolist,
-            tb.simple_broadcast(
-                tb.astensor(1), tb.astensor([2, 3, 4]), tb.astensor([5, 6, 7])
-            ),
+        == [[1, 1, 1], [2, 2, 2], [3, 3, 3]]
+    )
+    assert (
+        list(
+            map(
+                tb.tolist,
+                tb.simple_broadcast(
+                    tb.astensor(1), tb.astensor([2, 3, 4]), tb.astensor([5, 6, 7])
+                ),
+            )
         )
-    ) == [[1, 1, 1], [2, 3, 4], [5, 6, 7]]
-    assert list(
-        map(
-            tb.tolist,
-            tb.simple_broadcast(
-                tb.astensor([1]), tb.astensor([2, 3, 4]), tb.astensor([5, 6, 7])
-            ),
+        == [[1, 1, 1], [2, 3, 4], [5, 6, 7]]
+    )
+    assert (
+        list(
+            map(
+                tb.tolist,
+                tb.simple_broadcast(
+                    tb.astensor([1]), tb.astensor([2, 3, 4]), tb.astensor([5, 6, 7])
+                ),
+            )
         )
-    ) == [[1, 1, 1], [2, 3, 4], [5, 6, 7]]
+        == [[1, 1, 1], [2, 3, 4], [5, 6, 7]]
+    )
     with pytest.raises(Exception):
         tb.simple_broadcast(
[…]
