### 2020-10-13 fefa350
Apply suggestions from code review
Co-authored-by: Giordon Stark <kratsg@gmail.com>
