+++ README.rst
@@ -55,7 +55,6 @@ Alternatively the statistical model and observational data can be read from its 
 
    >>> import pyhf
    >>> import json
-   >>> wspace = pyhf.Workspace('likelihood.json')
    >>> wspace = pyhf.Workspace(json.load(open('test.json')))
    >>> model = wspace.model()
    >>> data = wspace.data(model)
