+++ README.rst
@@ -97,6 +97,10 @@ Finally, you can also use the command line interface that `pyhf` provides:
       ]
    }
    EOF
+
+which should produce the following JSON output:
+
+.. code:: bash
    {
       "CLs_exp": [
          0.0026062609501074576,
