### 2020-10-13 d033be3
Use consistent formatting everywhere
