### 2020-10-13 724e607
Combine to produce output
