+++ README.rst
@@ -54,14 +54,14 @@ Alternatively the statistical model and observational data can be read from its 
 .. code:: python
 
    >>> import pyhf
-   >>> import json
-   >>> wspace = pyhf.Workspace(json.load(open('likelihood.json')))
+   >>> import requests
+   >>> wspace = pyhf.Workspace(requests.get('https://git.io/JJYDE').json())
    >>> model = wspace.model()
    >>> data = wspace.data(model)
    >>> test_mu = 1.0
    >>> CLs_obs, CLs_exp = pyhf.infer.hypotest(test_mu, data, model, qtilde=True, return_expected=True)
    >>> print(f"Observed: {CLs_obs}, Expected: {CLs_exp}")
-   Observed: 0.05251497423736956, Expected: 0.06445320535890459
+   Observed: 0.3599840922126626, Expected: 0.3599840922126626
 
 
 Finally, you can also use the command line interface that `pyhf` provides:
