### 2020-10-13 db105eb
docs: Add JSON example to README (#1102)
* Add additional examples to the README that show structure of JSON spec
