### 2020-10-13 fddaa88
Make intervals available from import pyhf
