### 2020-10-13 d6b942e
Remove uneeded imports of intervals
