### 2020-10-13 fcdabd0
Remove unclear comment
