### 2020-10-13 64c9b23
Remove default arg for output-dir
