+++ tests/test_workspace.py
@@ -5,6 +5,7 @@ import pyhf.exceptions
 import json
 import logging
 import pyhf.workspace
+import pyhf.utils
 
 
 @pytest.fixture(
@@ -739,4 +740,6 @@ def test_closure_over_workspace_build():
     newdata = workspace.data(newmodel)
     two = pyhf.infer.hypotest(1.0, newdata, newmodel)
 
-    assert one == two
+    newworkspace = pyhf.Workspace.build(newmodel, newdata)
+
+    assert pyhf.utils.digest(newworkspace) == pyhf.utils.digest(workspace)
