+++ tests/test_paramsets.py
@@ -11,7 +11,7 @@ def test_paramset_unconstrained():
     )
     assert pset.suggested_init == [0, 1, 2, 3, 4]
     assert pset.suggested_bounds == [(-1, 1), (-2, 2), (-3, 3), (-4, 4)]
-    assert not pset.fixed
+    assert not pset.suggested_fixed
     assert not pset.constrained
 
 
@@ -26,7 +26,7 @@ def test_paramset_constrained_custom_sigmas():
     )
     assert pset.suggested_init == [0, 1, 2, 3, 4]
     assert pset.suggested_bounds == [(-1, 1), (-2, 2), (-3, 3), (-4, 4)]
-    assert not pset.fixed
+    assert not pset.suggested_fixed
     assert pset.constrained
     assert pset.width() == [1, 2, 3, 4, 5]
 
@@ -41,7 +41,7 @@ def test_paramset_constrained_default_sigmas():
     )
     assert pset.suggested_init == [0, 1, 2, 3, 4]
     assert pset.suggested_bounds == [(-1, 1), (-2, 2), (-3, 3), (-4, 4)]
-    assert not pset.fixed
+    assert not pset.suggested_fixed
     assert pset.constrained
     assert pset.width() == [1, 1, 1, 1, 1]
 
@@ -57,7 +57,7 @@ def test_paramset_constrained_custom_factors():
     )
     assert pset.suggested_init == [0, 1, 2, 3, 4]
     assert pset.suggested_bounds == [(-1, 1), (-2, 2), (-3, 3), (-4, 4)]
-    assert not pset.fixed
+    assert not pset.suggested_fixed
     assert pset.constrained
     assert pset.width() == [1 / 10.0, 1 / 20.0, 1 / 30.0, 1 / 40.0, 1 / 50.0]
 
