### 2020-10-14 f951465
Update src/pyhf/workspace.py
Co-authored-by: Matthew Feickert <matthew.feickert@cern.ch>
