### 2020-10-14 194baa0
Update src/pyhf/exceptions/__init__.py
Co-authored-by: Matthew Feickert <matthew.feickert@cern.ch>
