+++ CONTRIBUTING.md
@@ -6,7 +6,7 @@ To get started fork the repo.
 ## Issues
 
 Making Issues is very helpful to the project &mdash; they help the dev team form the development roadmap and are where most important discussion takes place.
-If you have suggestions, questions that you can't find answers to in the documentation or on the Stack Overflow tag, or have found a bug please open an Issue!
+If you have suggestions, questions that you can't find answers to in the [documentation website](https://scikit-hep.org/pyhf/) or on the [Stack Overflow tag](https://stackoverflow.com/questions/tagged/pyhf), or have found a bug please [open an Issue](https://github.com/scikit-hep/pyhf/issues/new/choose)!
 
 ## Pull Requests
 
