+++ CONTRIBUTING.md
@@ -23,9 +23,9 @@ If you're looking for some examples of high quality contributed pull requests we
 
 Many thanks goes out to our contributors!
 
-### WIP
+### Drafts
 
-Unless you are making a single commit pull request please create a WIP pull request. Outline the work that will be done in this ongoing pull request. When you are close to being done please assign someone with Approver permissions to follow the pull request.
+Unless you are making a single commit pull request please create a draft pull request. Outline the work that will be done in this ongoing pull request. When you are close to being done please tag someone with Approver permissions to follow the pull request.
 
 ### Pull Requests Procedure
 
