### 2020-10-14 6d4eb4e
feat: Add pyhf.workspace.build (#1101)
* Add workspace class method to build workspaces from a model and data
* Add test to validate build behavior
