+++ src/pyhf/exceptions/__init__.py
@@ -22,9 +22,10 @@ class InvalidSpecification(Exception):
     InvalidSpecification is raised when a specification does not validate against the given schema.
     """
 
-    def __init__(self, ValidationError):
+    def __init__(self, ValidationError, schema=None):
         self.exc_info = sys.exc_info()
         self.parent = ValidationError
+        self.schema = schema
         self.path = ''
         for item in ValidationError.path:
             if isinstance(item, int):
@@ -33,9 +34,7 @@ class InvalidSpecification(Exception):
                 self.path += '.{}'.format(item)
         self.path = self.path.lstrip('.')
         self.instance = ValidationError.instance
-        message = '{0}.\n\tPath: {1}\n\tInstance: {2}'.format(
-            ValidationError.message, self.path, self.instance
-        )
+        message = f"{ValidationError.message}.\n\tPath: {self.path}\n\tInstance: {self.instance} Schema: {self.schema}"
         # Call the base class constructor with the parameters it needs
         super().__init__(message)
 
