### 2020-10-15 ace6d8f
refactor: Rename paramset.fixed to paramset.suggested_fixed (#1110)
* Rename paramset.fixed to paramset.suggested_fixed to harmonize the API more
