+++ src/pyhf/infer/__init__.py
@@ -11,7 +11,7 @@ def hypotest(
     init_pars=None,
     par_bounds=None,
     fixed_params=None,
-    qtilde=False,
+    qtilde=True,
     **kwargs,
 ):
     r"""
