### 2020-10-15 abe4ebd
docs: Add importance of Issues to CONTRIBUTING.md (#1113)
* Note importance of Issues to communicating with the dev team
* Update terminology from WIP to Drafts
* Clarify pull request procedure
