### 2020-10-15 1006ab4
add test for workspace merge/combine
