### 2020-10-15 587b782
add ability to change default backend
