+++ src/new_pyhf/workspace.py
@@ -1,0 +1,678 @@
+"""
+pyhf workspaces hold the three data items:
+
+* the statistical model p(data|parameters)
+* the observed data (optional)
+* fit configurations ("measurements")
+"""
+import logging
+import jsonpatch
+import copy
+import collections
+from . import exceptions
+from . import utils
+from .pdf import Model
+from .mixins import _ChannelSummaryMixin
+
+log = logging.getLogger(__name__)
+
+
+def _join_items(join, left_items, right_items, key='name'):
+    """
+    Join two lists of dictionaries along the given key.
+
+    This is meant to be as generic as possible for any pairs of lists of dictionaries for many join operations.
+
+    Args:
+        join (`str`): The join operation to apply. See ~pyhf.workspace.Workspace for valid join operations.
+        left_items (`list`): A list of dictionaries to join on the left
+        right_items (`list`): A list of dictionaries to join on the right
+
+    Returns:
+        :obj:`list`: A joined list of dictionaries.
+
+    """
+    if join == 'right outer':
+        primary_items, secondary_items = right_items, left_items
+    else:
+        primary_items, secondary_items = left_items, right_items
+    joined_items = copy.deepcopy(primary_items)
+    for secondary_item in secondary_items:
+        # outer join: merge primary and secondary, matching where possible
+        if join == 'outer' and secondary_item in primary_items:
+            continue
+        # left/right outer join: only add secondary if existing item (by key value) is not in primary
+        # NB: this will be slow for large numbers of items
+        elif join in ['left outer', 'right outer'] and secondary_item[key] in [
+            item[key] for item in joined_items
+        ]:
+            continue
+        joined_items.append(copy.deepcopy(secondary_item))
+    return joined_items
+
+
+def _join_versions(join, left_version, right_version):
+    """
+    Join two workspace versions.
+
+    Raises:
+      ~pyhf.exceptions.InvalidWorkspaceOperation: Versions are incompatible.
+
+    Args:
+        join (`str`): The join operation to apply. See ~pyhf.workspace.Workspace for valid join operations.
+        left_version (`str`): The left workspace version.
+        right_version (`str`): The right workspace version.
+
+    Returns:
+        :obj:`str`: The workspace version.
+
+    """
+    if left_version != right_version:
+        raise exceptions.InvalidWorkspaceOperation(
+            f"Workspaces of different versions cannot be combined: {left_version} != {right_version}"
+        )
+    return left_version
+
+
+def _join_channels(join, left_channels, right_channels):
+    """
+    Join two workspace channel specifications.
+
+    Raises:
+      ~pyhf.exceptions.InvalidWorkspaceOperation: Channel specifications are incompatible.
+
+    Args:
+        join (`str`): The join operation to apply. See ~pyhf.workspace.Workspace for valid join operations.
+        left_channels (`list`): The left channel specification.
+        right_channels (`list`): The right channel specification.
+
+    Returns:
+        :obj:`list`: A joined list of channels. Each channel follows the :obj:`defs.json#/definitions/channel` `schema <https://scikit-hep.org/pyhf/likelihood.html#channel>`__
+
+    """
+
+    joined_channels = _join_items(join, left_channels, right_channels)
+    if join == 'none':
+        common_channels = set(c['name'] for c in left_channels).intersection(
+            c['name'] for c in right_channels
+        )
+        if common_channels:
+            raise exceptions.InvalidWorkspaceOperation(
+                f"Workspaces cannot have any channels in common with the same name: {common_channels}. You can also try a different join operation: {Workspace.valid_joins}."
+            )
+
+    elif join == 'outer':
+        counted_channels = collections.Counter(
+            channel['name'] for channel in joined_channels
+        )
[…]
