+++ src/pyhf/__init__.py
@@ -1,169 +1,221 @@
-from .tensor import BackendRetriever as tensor
-from .optimize import OptimizerRetriever as optimize
-from .version import __version__
-from .exceptions import InvalidBackend, InvalidOptimizer, Unsupported
-from . import events
+import logging
 
-tensorlib = None
-optimizer = None
+from .. import exceptions
+from .. import get_backend
+
+log = logging.getLogger(__name__)
+
+registry = {}
 
 
-def get_backend():
+def validate_modifier_structure(modifier):
     """
-    Get the current backend and the associated optimizer
-
-    Example:
-        >>> import pyhf
-        >>> backend, optimizer = pyhf.get_backend()
-        >>> backend
-        <pyhf.tensor.numpy_backend.numpy_backend object at 0x...>
-        >>> optimizer
-        <pyhf.optimize.scipy_optimizer object at 0x...>
-
-    Returns:
-        backend, optimizer
+    Check if given object contains the right structure for modifiers
     """
-    global tensorlib
-    global optimizer
-    return tensorlib, optimizer
+    required_methods = ['required_parset']
+
+    for method in required_methods:
+        if not hasattr(modifier, method):
+            raise exceptions.InvalidModifier(
+                'Expected {0:s} method on modifier {1:s}'.format(
+                    method, modifier.__name__
+                )
+            )
+    return True
 
 
-tensorlib = tensor.numpy_backend()
-default_backend = tensorlib
-optimizer = optimize.scipy_optimizer()
-default_optimizer = optimizer
-
-
-@events.register('change_backend')
-def set_backend(backend, custom_optimizer=None, precision=None):
+def add_to_registry(
+    cls, cls_name=None, constrained=False, pdf_type='normal', op_code='addition'
+):
     """
-    Set the backend and the associated optimizer
+    Consistent add_to_registry() function that handles actually adding thing to the registry.
 
-    Example:
-        >>> import pyhf
-        >>> pyhf.set_backend("tensorflow")
-        >>> pyhf.tensorlib.name
-        'tensorflow'
-        >>> pyhf.tensorlib.precision
-        '32b'
-        >>> pyhf.set_backend(b"pytorch", precision="64b")
-        >>> pyhf.tensorlib.name
-        'pytorch'
-        >>> pyhf.tensorlib.precision
-        '64b'
-        >>> pyhf.set_backend(pyhf.tensor.numpy_backend())
-        >>> pyhf.tensorlib.name
-        'numpy'
-        >>> pyhf.tensorlib.precision
-        '64b'
+    Raises an error if the name to register for the modifier already exists in the registry,
+    or if the modifier does not have the right structure.
+    """
+    global registry
+    cls_name = cls_name or cls.__name__
+    if cls_name in registry:
+        raise KeyError('The modifier name "{0:s}" is already taken.'.format(cls_name))
+    # validate the structure
+    validate_modifier_structure(cls)
+    # set is_constrained
+    cls.is_constrained = constrained
+    if constrained:
+        tensorlib, _ = get_backend()
+        if not hasattr(tensorlib, pdf_type):
+            raise exceptions.InvalidModifier(
+                'The specified pdf_type "{0:s}" is not valid for {1:s}({2:s}). See pyhf.tensor documentation for available pdfs.'.format(
+                    pdf_type, cls_name, cls.__name__
+                )
+            )
+        cls.pdf_type = pdf_type
+    else:
+        cls.pdf_type = None
+
+    if op_code not in ['addition', 'multiplication']:
+        raise exceptions.InvalidModifier(
+            'The specified op_code "{0:s}" is not valid for {1:s}({2:s}). See pyhf.modifier documentation for available operation codes.'.format(
+                op_code, cls_name, cls.__name__
+            )
+        )
[…]
