+++ src/new_pyhf/optimize/common.py
@@ -1,154 +1,0 @@
-"""Common Backend Shim to prepare minimization for optimizer."""
-from .. import get_backend
-from ..tensor.common import _TensorViewer
-
-
-def _make_stitch_pars(tv=None, fixed_values=None):
-    """
-    Construct a callable to stitch fixed paramter values into the unfixed parameters. See :func:`shim`.
-
-    This is extracted out to be unit-tested for proper behavior.
-
-    If ``tv`` or ``fixed_values`` are not provided, this returns the identity callable.
-
-    Args:
-        tv (~pyhf.tensor.common._TensorViewer): tensor viewer instance
-        fixed_values (`list`): default set of values to stitch parameters with
-
-    Returns:
-        callable (`func`): a callable that takes nuisance parameter values as input
-    """
-    if tv is None or fixed_values is None:
-        return lambda pars, stitch_with=None: pars
-
-    def stitch_pars(pars, stitch_with=fixed_values):
-        tb, _ = get_backend()
-        return tv.stitch([tb.astensor(stitch_with, dtype='float'), pars])
-
-    return stitch_pars
-
-
-def _get_tensor_shim():
-    """
-    A shim-retriever to lazy-retrieve the necessary shims as needed.
-
-    Because pyhf.tensor is a lazy-retriever for the backends, we can be sure
-    that tensorlib is imported correctly.
-    """
-    tensorlib, _ = get_backend()
-    if tensorlib.name == 'numpy':
-        from .opt_numpy import wrap_objective as numpy_shim
-
-        return numpy_shim
-
-    if tensorlib.name == 'tensorflow':
-        from .opt_tflow import wrap_objective as tflow_shim
-
-        return tflow_shim
-
-    if tensorlib.name == 'pytorch':
-        from .opt_pytorch import wrap_objective as pytorch_shim
-
-        return pytorch_shim
-
-    if tensorlib.name == 'jax':
-        from .opt_jax import wrap_objective as jax_shim
-
-        return jax_shim
-    raise ValueError(f'No optimizer shim for {tensorlib.name}.')
-
-
-def shim(
-    objective,
-    data,
-    pdf,
-    init_pars,
-    par_bounds,
-    fixed_vals=None,
-    do_grad=False,
-    do_stitch=False,
-):
-    """
-    Prepare Minimization for Optimizer.
-
-    Args:
-        objective (`func`): objective function
-        data (`list`): observed data
-        pdf (~pyhf.pdf.Model): The statistical model adhering to the schema model.json
-        init_pars (`list`): initial parameters
-        par_bounds (`list`): parameter boundaries
-        fixed_vals (`list`): fixed parameter values
-
-    .. note::
-
-        ``minimizer_kwargs`` is a dictionary containing
-
-          - ``func`` (`func`): backend-wrapped ``objective`` function (potentially with gradient)
-          - ``x0`` (`list`):  modified initializations for minimizer
-          - ``do_grad`` (`bool`): whether or not gradient is used
-          - ``bounds`` (`list`): modified bounds for minimizer
-          - ``fixed_vals`` (`list`): modified fixed values for minimizer
-
-    .. note::
-
-        ``stitch_pars(pars, stitch_with=None)`` is a callable that will
-        stitch the fixed parameters of the minimization back into the unfixed
-        parameters.
-
-    .. note::
-
-        ``do_stitch`` will modify the ``init_pars``, ``par_bounds``, and ``fixed_vals`` by stripping away the entries associated with fixed parameters. The parameters can be stitched back in via ``stitch_pars``.
-
-    Returns:
[…]
