+++ src/new_pyhf/version.py
@@ -1,5 +1,0 @@
-"""Define pyhf version information."""
-
-# Use semantic versioning (https://semver.org/)
-# The version number is controlled through bumpversion.cfg
-__version__ = '0.5.0'
