+++ src/new_pyhf/contrib/viz/__init__.py
@@ -1,1 +1,0 @@
-"""Visualizations of pyhf models and results."""

[…]
