### 2020-10-15 37992b0
Apply suggestions from code review
Co-authored-by: Matthew Feickert <matthew.feickert@cern.ch>
