### 2020-10-15 #2734 @eloisabentivegna
> I don't think there's any conflict of interest, but I'll just note here to be transparent that @betatim knows the authors quite well.

Thanks for letting me know, @matthewfeickert. @betatim, if you are available to review, could you first check out the conflict-of-interest policy (https://joss.readthedocs.io/en/latest/reviewer_guidelines.html#joss-conflict-of-interest-policy)?
