+++ src/pyhf/__init__.py
@@ -1,221 +1,190 @@
-import logging
+from .tensor import BackendRetriever as tensor
+from .optimize import OptimizerRetriever as optimize
+from .version import __version__
+from .exceptions import InvalidBackend, InvalidOptimizer, Unsupported
+from . import events
 
-from .. import exceptions
-from .. import get_backend
+tensorlib = None
+optimizer = None
+default_backend = None
+default_optimizer = None
 
-log = logging.getLogger(__name__)
-
-registry = {}
-
-
-def validate_modifier_structure(modifier):
+def get_backend():
     """
-    Check if given object contains the right structure for modifiers
-    """
-    required_methods = ['required_parset']
+    Get the current backend and the associated optimizer
 
-    for method in required_methods:
-        if not hasattr(modifier, method):
-            raise exceptions.InvalidModifier(
-                'Expected {0:s} method on modifier {1:s}'.format(
-                    method, modifier.__name__
-                )
-            )
-    return True
-
-
-def add_to_registry(
-    cls, cls_name=None, constrained=False, pdf_type='normal', op_code='addition'
-):
-    """
-    Consistent add_to_registry() function that handles actually adding thing to the registry.
-
-    Raises an error if the name to register for the modifier already exists in the registry,
-    or if the modifier does not have the right structure.
-    """
-    global registry
-    cls_name = cls_name or cls.__name__
-    if cls_name in registry:
-        raise KeyError('The modifier name "{0:s}" is already taken.'.format(cls_name))
-    # validate the structure
-    validate_modifier_structure(cls)
-    # set is_constrained
-    cls.is_constrained = constrained
-    if constrained:
-        tensorlib, _ = get_backend()
-        if not hasattr(tensorlib, pdf_type):
-            raise exceptions.InvalidModifier(
-                'The specified pdf_type "{0:s}" is not valid for {1:s}({2:s}). See pyhf.tensor documentation for available pdfs.'.format(
-                    pdf_type, cls_name, cls.__name__
-                )
-            )
-        cls.pdf_type = pdf_type
-    else:
-        cls.pdf_type = None
-
-    if op_code not in ['addition', 'multiplication']:
-        raise exceptions.InvalidModifier(
-            'The specified op_code "{0:s}" is not valid for {1:s}({2:s}). See pyhf.modifier documentation for available operation codes.'.format(
-                op_code, cls_name, cls.__name__
-            )
-        )
-    cls.op_code = op_code
-
-    registry[cls_name] = cls
-
-
-def modifier(*args, **kwargs):
-    """
-    Decorator for registering modifiers. To flag the modifier as a constrained modifier, add `constrained=True`.
-
-
-    Args:
-        name: the name of the modifier to use. Use the class name by default. (default: None)
-        constrained: whether the modifier is constrained or not. (default: False)
-        pdf_type: the name of the pdf to use from tensorlib if constrained. (default: normal)
-        op_code: the name of the operation the modifier performs on the data (e.g. addition, multiplication)
+    Example:
+        >>> import pyhf
+        >>> backend, optimizer = pyhf.get_backend()
+        >>> backend
+        <pyhf.tensor.numpy_backend.numpy_backend object at 0x...>
+        >>> optimizer
+        <pyhf.optimize.scipy_optimizer object at 0x...>
 
     Returns:
-        modifier
-
-    Raises:
-        ValueError: too many keyword arguments, or too many arguments, or wrong arguments
-        TypeError: provided name is not a string
[…]
