+++ src/pyhf/cli/__init__.py
@@ -4,6 +4,6 @@ from .rootio import cli as rootio
 from .spec import cli as spec
 from .infer import cli as infer
 from .complete import cli as complete
-from ..contrib.cli import cli as contrib
+from ..contrib import cli as contrib
 
 __all__ = ['cli', 'rootio', 'spec', 'infer', 'complete', 'contrib']
