### 2020-10-15 db3fbc7
Remove uneeded env removals
