+++ tests/test_scripts.py
@@ -629,6 +629,15 @@ def test_missing_contrib_download(caplog):
         else:
             import_module("pyhf.cli")
 
+        # Force environment for runner
+        for module in [
+            "pyhf.contrib",
+            "pyhf.contrib.cli",
+            "pyhf.contrib.utils",
+        ]:
+            if module in sys.modules:
+                del sys.modules[module]
+
         from pyhf.contrib.cli import download
 
         runner = CliRunner(mix_stderr=False)
