+++ tests/test_scripts.py
@@ -631,14 +631,14 @@ def test_missing_contrib_download(caplog):
 
         # Force environment for runner
         for module in [
-            "pyhf.cli.contrib",
             "pyhf.contrib",
+            "pyhf.contrib.cli",
             "pyhf.contrib.utils",
         ]:
             if module in sys.modules:
                 del sys.modules[module]
 
-        from pyhf.cli.contrib import download
+        from pyhf.contrib.cli import download
 
         runner = CliRunner(mix_stderr=False)
         result = runner.invoke(
