### 2020-10-16 6516e0a
clean up _join_items logic more
