+++ README.rst
@@ -8,9 +8,8 @@ pure-python fitting/limit-setting/interval estimation HistFactory-style
 
 |GitHub Project| |DOI| |Scikit-HEP| |NSF Award Number|
 
-|GitHub Actions Status: CI| |GitHub Actions Status: Publish| |Docker
-Automated| |Code Coverage| |Language grade: Python| |CodeFactor| |Code
-style: black|
+|GitHub Actions Status: CI| |GitHub Actions Status: Docs| |GitHub Actions Status: Publish|
+|Docker Automated| |Code Coverage| |Language grade: Python| |CodeFactor| |Code style: black|
 
 |Docs| |Binder|
 
@@ -318,6 +317,8 @@ and grant `OAC-1450377 <https://www.nsf.gov/awardsearch/showAward?AWD_ID=1450377
    :target: https://nsf.gov/awardsearch/showAward?AWD_ID=1836650
 .. |GitHub Actions Status: CI| image:: https://github.com/scikit-hep/pyhf/workflows/CI/CD/badge.svg
    :target: https://github.com/scikit-hep/pyhf/actions?query=workflow%3ACI%2FCD+branch%3Amaster
+.. |GitHub Actions Status: Docs| image:: https://github.com/scikit-hep/pyhf/workflows/Docs/badge.svg?branch=master
+   :target: https://github.com/scikit-hep/pyhf/actions?query=workflow%3ADocs+branch%3Amaster
 .. |GitHub Actions Status: Publish| image:: https://github.com/scikit-hep/pyhf/workflows/publish%20distributions/badge.svg
    :target: https://github.com/scikit-hep/pyhf/actions?query=workflow%3A%22publish+distributions%22+branch%3Amaster
 .. |Docker Automated| image:: https://img.shields.io/docker/automated/pyhf/pyhf.svg
