### 2020-10-16 b7038ea
ci: Make docs separate workflow (#1120)
* Migrate the testing and deploying of the docs into a separate GitHub Actions workflow
   - Allows for running independently using workflow dispatch
* Add Docs build status on master badge to README
