+++ .github/workflows/ci.yml
@@ -46,64 +46,6 @@ jobs:
       run: |
         python -m pytest -r sx --benchmark-sort=mean tests/benchmarks/test_benchmark.py
 
-  docs:
-
-    runs-on: ubuntu-latest
-    strategy:
-      matrix:
-        python-version: [3.8]
-
-    steps:
-    - uses: actions/checkout@v2
-      with:
-        fetch-depth: 0
-    - name: Set up Python 3.8
-      uses: actions/setup-python@v2
-      with:
-        python-version: ${{ matrix.python-version }}
-    - name: Install dependencies
-      run: |
-        python -m pip install --upgrade pip setuptools wheel
-        python -m pip install --ignore-installed -U -q --no-cache-dir -e .[docs,test]
-        python -m pip list
-        sudo apt-get update
-        sudo apt-get -qq install pandoc
-    - name: Check docstrings
-      run: |
-        # Group 1 is related to docstrings
-        pydocstyle --select D1 src/pyhf/pdf.py \
-                               src/pyhf/workspace.py \
-                               src/pyhf/probability.py \
-                               src/pyhf/patchset.py \
-                               src/pyhf/interpolators \
-                               src/pyhf/infer \
-                               src/pyhf/optimize \
-                               src/pyhf/contrib \
-                               src/pyhf/cli
-    - name: Test and build docs
-      run: |
-        python -m doctest README.rst
-        python setup.py build_sphinx
-        touch docs/_build/html/.nojekyll
-    - name: Check schemas are copied over
-      run: |
-        # is a directory
-        [ -d "docs/_build/html/schemas" ]
-        # is not a symlink
-        [ ! -L "docs/_build/html/schemas" ]
-        # is not empty
-        [ "$(ls -A docs/_build/html/schemas)" ]
-    - name: Deploy docs to GitHub Pages
-      if: success() && github.event_name == 'push' && github.ref == 'refs/heads/master'
-      uses: peaceiris/actions-gh-pages@v3
-      with:
-        github_token: ${{ secrets.GITHUB_TOKEN }}
-        publish_dir: docs/_build/html
-        force_orphan: true
-        user_name: 'github-actions[bot]'
-        user_email: 'github-actions[bot]@users.noreply.github.com'
-        commit_message: Deploy to GitHub pages
-
   docker:
 
     runs-on: ubuntu-latest
