### 2020-10-16 6470874
add default arg to get_backend, change instances of default_backend
