+++ src/pyhf/interpolators/code1.py
@@ -1,7 +1,7 @@
 """Piecewise-Exponential Interpolation (Code 1)."""
 import logging
 import math
-from .. import get_backend, default_backend
+from .. import get_backend
 from .. import events
 from . import _slow_interpolator_looper
 
@@ -26,6 +26,7 @@ class code1:
 
     def __init__(self, histogramssets, subscribe=True):
         """Piecewise-Exponential Interpolation."""
+        default_backend, _ = get_backend(default=True)
         # nb: this should never be a tensor, store in default backend (e.g. numpy)
         self._histogramssets = default_backend.astensor(histogramssets)
         # initial shape will be (nsysts, 1)
