+++ src/pyhf/interpolators/code4p.py
@@ -1,6 +1,6 @@
 """Piecewise-Linear + Polynomial Interpolation (Code 4p)."""
 import logging
-from .. import get_backend, default_backend
+from .. import get_backend
 from .. import events
 from . import _slow_interpolator_looper
 
@@ -18,6 +18,7 @@ class code4p:
 
     def __init__(self, histogramssets, subscribe=True):
         """Piecewise-Linear  + Polynomial Interpolation."""
+        default_backend, _ = get_backend(default=True)
         # nb: this should never be a tensor, store in default backend (e.g. numpy)
         self._histogramssets = default_backend.astensor(histogramssets)
         # initial shape will be (nsysts, 1)
