+++ src/pyhf/parameters/paramview.py
@@ -44,7 +44,7 @@ class ParamViewer:
 
     def __init__(self, shape, par_map, par_selection):
         default_backend, _ = get_backend(default=True)
-        
+
         batch_size = shape[0] if len(shape) > 1 else None
 
         fullsize = default_backend.product(default_backend.astensor(shape))
