### 2020-10-16 697ea60
feat: Merge channel functionality in pyhf combine (#1088)
* Extend combine to allow for merging channels from two workspaces when performing an outer join
  - This is generally useful when you want to combine a bkg-only workspace with a signal-only workspace with the same set of channels, but need to merge the channels together (extend bkg-only samples with signal-only samples for each channel)
* Improve the coverage of pyhf/workspace.py by cleaning up the pyhf.workspace._join_items logic
