+++ tests/test_workspace.py
@@ -6,6 +6,7 @@ import json
 import logging
 import pyhf.workspace
 import pyhf.utils
+import copy
 
 
 @pytest.fixture(
@@ -140,10 +141,14 @@ def test_workspace_observations(workspace_factory):
     assert w.observations
 
 
-def test_get_workspace_data(workspace_factory):
+@pytest.mark.parametrize(
+    "with_aux",
+    [True, False],
+)
+def test_get_workspace_data(workspace_factory, with_aux):
     w = workspace_factory()
     m = w.model()
-    assert w.data(m)
+    assert w.data(m, with_aux=with_aux)
 
 
 def test_get_workspace_data_bad_model(workspace_factory, caplog):
@@ -300,8 +305,14 @@ def test_rename_measurement(workspace_factory):
 
 @pytest.fixture(scope='session')
 def join_items():
-    left = [{'name': 'left', 'key': 'value'}, {'name': 'common', 'key': 'left'}]
-    right = [{'name': 'right', 'key': 'value'}, {'name': 'common', 'key': 'right'}]
+    left = [
+        {'name': 'left', 'key': 'value', 'deep': [{'name': 1}]},
+        {'name': 'common', 'key': 'left', 'deep': [{'name': 1}]},
+    ]
+    right = [
+        {'name': 'right', 'key': 'value', 'deep': [{'name': 2}]},
+        {'name': 'common', 'key': 'right', 'deep': [{'name': 2}]},
+    ]
     return (left, right)
 
 
@@ -335,6 +346,28 @@ def test_join_items_right_outer(join_items):
     )
     assert not all(left in joined for left in left_items)
     assert all(right in joined for right in right_items)
+
+
+def test_join_items_left_outer_deep(join_items):
+    left_items, right_items = join_items
+    joined = pyhf.workspace._join_items(
+        'left outer', left_items, right_items, key='name', deep_merge_key='deep'
+    )
+    assert [k['deep'] for k in joined if k['name'] == 'common'][0] == [
+        {'name': 1},
+        {'name': 2},
+    ]
+
+
+def test_join_items_right_outer_deep(join_items):
+    left_items, right_items = join_items
+    joined = pyhf.workspace._join_items(
+        'right outer', left_items, right_items, key='name', deep_merge_key='deep'
+    )
+    assert [k['deep'] for k in joined if k['name'] == 'common'][0] == [
+        {'name': 2},
+        {'name': 1},
+    ]
 
 
 @pytest.mark.parametrize("join", ['none', 'outer'])
@@ -508,6 +541,17 @@ def test_combine_workspace_invalid_join_operation(workspace_factory, join):
 
 
 @pytest.mark.parametrize("join", ['none'])
+def test_combine_workspace_invalid_join_operation_merge(workspace_factory, join):
+    ws = workspace_factory()
+    new_ws = ws.rename(
+        channels={channel: f'renamed_{channel}' for channel in ws.channels},
+    )
+    with pytest.raises(ValueError) as excinfo:
+        pyhf.Workspace.combine(ws, new_ws, join=join, merge_channels=True)
+    assert join in str(excinfo.value)
+
+
+@pytest.mark.parametrize("join", ['none'])
 def test_combine_workspace_incompatible_parameter_configs(workspace_factory, join):
     ws = workspace_factory()
     new_ws = ws.rename(
@@ -541,6 +585,36 @@ def test_combine_workspace_incompatible_parameter_configs_outer_join(
     assert new_ws.get_measurement(measurement_name='GaussExample')['config'][
         'parameters'
     ][0]['name'] in str(excinfo.value)
+
+
+@pytest.mark.parametrize("join", ['outer'])
+def test_combine_workspace_compatible_parameter_configs_outer_join(
[…]
