### 2020-10-16 f248db0
docs: Use sphinx obj scope in docs by default (#1122)
* Harmonize use of sphinx :obj: notation in docs
