+++ src/pyhf/tensor/jax_backend.py
@@ -74,9 +74,9 @@ class jax_backend(object):
             DeviceArray([-1., -1.,  0.,  1.,  1.], dtype=float64)
 
         Args:
-            tensor_in (`tensor`): The input tensor object
-            min_value (`scalar` or `tensor` or `None`): The minimum value to be cliped to
-            max_value (`scalar` or `tensor` or `None`): The maximum value to be cliped to
+            tensor_in (:obj:`tensor`): The input tensor object
+            min_value (:obj:`scalar` or :obj:`tensor` or :obj:`None`): The minimum value to be cliped to
+            max_value (:obj:`scalar` or :obj:`tensor` or :obj:`None`): The maximum value to be cliped to
 
         Returns:
             JAX ndarray: A clipped `tensor`
@@ -97,7 +97,7 @@ class jax_backend(object):
                           0.99532227], dtype=float64)
 
         Args:
-            tensor_in (`tensor`): The input tensor object
+            tensor_in (:obj:`tensor`): The input tensor object
 
         Returns:
             JAX ndarray: The values of the error function at the given points.
@@ -117,7 +117,7 @@ class jax_backend(object):
             DeviceArray([-2., -1.,  0.,  1.,  2.], dtype=float64)
 
         Args:
-            tensor_in (`tensor`): The input tensor object
+            tensor_in (:obj:`tensor`): The input tensor object
 
         Returns:
             JAX ndarray: The values of the inverse of the error function at the given points.
@@ -138,8 +138,8 @@ class jax_backend(object):
                          [2., 2.]], dtype=float64)
 
         Args:
-            tensor_in (`Tensor`): The tensor to be repeated
-            repeats (`Tensor`): The tuple of multipliers for each dimension
+            tensor_in (:obj:`tensor`): The tensor to be repeated
+            repeats (:obj:`tensor`): The tuple of multipliers for each dimension
 
         Returns:
             JAX ndarray: The tensor with repeated axes
@@ -161,9 +161,9 @@ class jax_backend(object):
             DeviceArray([9.], dtype=float64)
 
         Args:
-            predicate (`scalar`): The logical condition that determines which callable to evaluate
-            true_callable (`callable`): The callable that is evaluated when the :code:`predicate` evalutes to :code:`true`
-            false_callable (`callable`): The callable that is evaluated when the :code:`predicate` evalutes to :code:`false`
+            predicate (:obj:`scalar`): The logical condition that determines which callable to evaluate
+            true_callable (:obj:`callable`): The callable that is evaluated when the :code:`predicate` evalutes to :code:`true`
+            false_callable (:obj:`callable`): The callable that is evaluated when the :code:`predicate` evalutes to :code:`false`
 
         Returns:
             JAX ndarray: The output of the callable that was evaluated
@@ -340,9 +340,9 @@ class jax_backend(object):
             DeviceArray([0.16062314, 0.12407692], dtype=float64)
 
         Args:
-            n (`tensor` or `float`): The value at which to evaluate the approximation to the Poisson distribution p.m.f.
+            n (:obj:`tensor` or :obj:`float`): The value at which to evaluate the approximation to the Poisson distribution p.m.f.
                                   (the observed number of events)
-            lam (`tensor` or `float`): The mean of the Poisson distribution p.m.f.
+            lam (:obj:`tensor` or :obj:`float`): The mean of the Poisson distribution p.m.f.
                                     (the expected number of events)
 
         Returns:
@@ -384,9 +384,9 @@ class jax_backend(object):
             DeviceArray([0.35206533, 0.46481887], dtype=float64)
 
         Args:
-            x (`tensor` or `float`): The value at which to evaluate the Normal distribution p.d.f.
[…]
