### 2020-10-16 772768a
Wrap func to get set_backend docs
