### 2020-10-17 5e88cf8
fix: Use functools.wraps with events.register to render docstrings (#1124)
* Use functools.wraps to wrap events.register decorator to properly get docstrings for pyhf.set_backend
