### 2020-10-17 a62657d
add utility to remove prefix string
