### 2020-10-17 cd61460
update writexml to add prefix
