### 2020-10-17 3469535
fix up test with new build_measurement signature
