### 2020-10-17 38d7a93
validate inputs on hypotest
