+++ tests/test_infer.py
@@ -77,6 +77,22 @@ def test_hypotest_default(tmpdir, hypotest_args):
     assert isinstance(result, type(tb.astensor(result)))
 
 
+def test_hypotest_poi_outofbounds(tmpdir, hypotest_args):
+    """
+    Check that the default return structure of pyhf.infer.hypotest is as expected
+    """
+    pdf = pyhf.simplemodels.hepdata_like(
+        signal_data=[12.0, 11.0], bkg_data=[50.0, 52.0], bkg_uncerts=[3.0, 7.0]
+    )
+    data = [51, 48] + pdf.config.auxdata
+
+    with pytest.raises(ValueError):
+        pyhf.infer.hypotest(-1.0, data, pdf)
+
+    with pytest.raises(ValueError):
+        pyhf.infer.hypotest(10.1, data, pdf)
+
+
 def test_hypotest_return_tail_probs(tmpdir, hypotest_args):
     """
     Check that the return structure of pyhf.infer.hypotest with the
