### 2020-10-17 6f30397
move check to fit
