### 2020-10-17 6f23ad4
move check to fit
