### 2020-10-17 4c5514e
move check to fit
