### 2020-10-18 8c2e86b
Expand test docstring
