### 2020-10-18 0748649
Add magic headers
