### 2020-10-19 bce96db
Add first check of compression type
