### 2020-10-19 ac42975
fix: Validate fit initialization parameters (#1126)
* Add check to validate initialization parameters for pyhf.infer.fit
* Add test of raised error using hypotest POI out of bounds
