### 2020-10-19 ef8c3e5
fix up coverage and add tests
