### 2020-10-19 5a743a2
fix: Logic bug for outer joins of merge channels (#1128)
* Fix logic bug for merge channels of combine for outer join functionality
* Amends PR #1088
