### 2020-10-19 eb72951
fix up tests
