+++ src/pyhf/readxml.py
@@ -223,6 +223,17 @@ def process_channel(channelxml, rootdir, track_progress=False):
 
 
 def process_measurements(toplvl, other_parameter_configs=None):
+    """
+    For a given XML structure, provide a parsed dictionary adhering to defs.json/#definitions/measurement.
+
+    Args:
+        toplvl (:module:`xml.etree.ElementTree`): The top-level XML document to parse.
+        other_parameter_configs (:obj:`list`): A list of other parameter configurations from other non-top-level XML documents to incorporate into the resulting measurement object.
+
+    Returns:
+        :obj:`dict`: A measurement object.
+
+    """
     results = []
     other_parameter_configs = other_parameter_configs if other_parameter_configs else []
 
