### 2020-10-20 5a930a7
gamma bins can't be individually set in pyhf so easily
