### 2020-10-20 ec2b9bb
fix: Fixed parameter naming for compatibility with ROOT import/export (#1125)
* Fix parameter naming for compatibility with ROOT import/export
  - prefixes alpha_ and gamma_ are stripped when importing from ROOT workspaces
  - prefixes alpha_ and gamma_ are added when exporting to ROOT workspaces
