+++ tests/test_import.py
@@ -47,11 +47,11 @@ def test_process_normfactor_configs():
     meas.append(poiel)
 
     setting = ET.Element('ParamSetting', Const='True')
-    setting.text = ' '.join(['Lumi', 'mu_both', 'mu_paramSettingOnly'])
+    setting.text = ' '.join(['Lumi', 'alpha_mu_both', 'alpha_mu_paramSettingOnly'])
     meas.append(setting)
 
     setting = ET.Element('ParamSetting', Val='2.0')
-    setting.text = ' '.join(['mu_both'])
+    setting.text = ' '.join(['alpha_mu_both'])
     meas.append(setting)
 
     toplvl.append(meas)
@@ -68,7 +68,7 @@ def test_process_normfactor_configs():
     meas.append(poiel)
 
     setting = ET.Element('ParamSetting', Val='3.0')
-    setting.text = ' '.join(['mu_both'])
+    setting.text = ' '.join(['alpha_mu_both'])
     meas.append(setting)
 
     toplvl.append(meas)
@@ -134,7 +134,7 @@ def test_import_measurements():
     assert 'parameters' in measurement_configs
     assert len(measurement_configs['parameters']) == 3
     parnames = [p['name'] for p in measurement_configs['parameters']]
-    assert sorted(parnames) == sorted(['lumi', 'SigXsecOverSM', 'alpha_syst1'])
+    assert sorted(parnames) == sorted(['lumi', 'SigXsecOverSM', 'syst1'])
 
     lumi_param_config = measurement_configs['parameters'][0]
     assert 'auxdata' in lumi_param_config
@@ -145,6 +145,34 @@ def test_import_measurements():
     assert lumi_param_config['inits'] == [1.0]
     assert 'sigmas' in lumi_param_config
     assert lumi_param_config['sigmas'] == [0.1]
+
+
+@pytest.mark.parametrize("const", ['False', 'True'])
+def test_import_measurement_gamma_bins(const):
+    toplvl = ET.Element("Combination")
+    meas = ET.Element(
+        "Measurement",
+        Name='NormalMeasurement',
+        Lumi=str(1.0),
+        LumiRelErr=str(0.017),
+        ExportOnly=str(True),
+    )
+    poiel = ET.Element('POI')
+    poiel.text = 'mu_SIG'
+    meas.append(poiel)
+
+    setting = ET.Element('ParamSetting', Const=const)
+    setting.text = ' '.join(['Lumi', 'alpha_mu_both', 'gamma_bin_0'])
+    meas.append(setting)
+
+    setting = ET.Element('ParamSetting', Val='2.0')
+    setting.text = ' '.join(['gamma_bin_0'])
+    meas.append(setting)
+
+    toplvl.append(meas)
+
+    with pytest.raises(ValueError):
+        pyhf.readxml.process_measurements(toplvl)
 
 
 def test_import_prepHistFactory():
