+++ src/pyhf/utils.py
@@ -111,3 +111,26 @@ def digest(obj, algorithm='sha256'):
             f"{algorithm} is not an algorithm provided by Python's hashlib library."
         )
     return hash_alg(stringified).hexdigest()
+
+
+def remove_prefix(text, prefix):
+    """
+    Remove a prefix from the beginning of the provided text.
+
+    Example:
+
+        >>> import pyhf
+        >>> pyhf.utils.remove_prefix("alpha_syst1", "alpha_")
+        'syst1'
+
+    Args:
+        text (:obj:`str`): A provided input to manipulate.
+        prefix (:obj:`str`): A prefix to remove from provided input, if it exists.
+
+    Returns:
+        stripped_text (:obj:`str`): Text with the prefix removed.
+    """
+    # NB: python3.9 can be `return text.removeprefix(prefix)`
+    if text.startswith(prefix):
+        return text[len(prefix) :]
+    return text
