+++ docs/bib/HEPData_likelihoods.bib
@@ -1,3 +1,12 @@
+@misc {1754675,
+    title = "{Search for squarks and gluinos in final states with same-sign leptons and jets using 139 fb$^{-1}$ of data collected with the ATLAS detector}",
+    doi = "10.17182/hepdata.91214.v3",
+    url = "https://doi.org/10.17182/hepdata.91214.v3",
+    year = "2020",
+    type = "Data Collection",
+    collaboration = "ATLAS",
+}
+
 @misc {ins1755298,
     title = "{Search for direct production of electroweakinos in final states with one lepton, missing transverse momentum and a Higgs boson decaying into two $b$-jets in (pp) collisions at $\sqrt{s}=13$ TeV with the ATLAS detector}",
     doi = "10.17182/hepdata.90607.v2",
