+++ docs/bib/HEPData_likelihoods.bib
@@ -1,4 +1,4 @@
-@misc {1754675,
+@misc {ins1754675,
     title = "{Search for squarks and gluinos in final states with same-sign leptons and jets using 139 fb$^{-1}$ of data collected with the ATLAS detector}",
     doi = "10.17182/hepdata.91214.v3",
     url = "https://doi.org/10.17182/hepdata.91214.v3",
