### 2020-10-20 ebcba8f
Use v0.2.4's style of scipy for osp_stats
