### 2020-10-20 1fb558f
Set new lowerbounds for JAX
