### 2020-10-20 9849858
Use jnp instead of np
