### 2020-10-20 e8147f1
Change np to jnp
