+++ src/pyhf/tensor/jax_backend.py
@@ -57,9 +57,9 @@ class jax_backend(object):
         self.name = 'jax'
         self.precision = kwargs.get('precision', '64b')
         self.dtypemap = {
-            'float': np.float64 if self.precision == '64b' else np.float32,
-            'int': np.int64 if self.precision == '64b' else np.int32,
-            'bool': np.bool_,
+            'float': jnp.float64 if self.precision == '64b' else jnp.float32,
+            'int': jnp.int64 if self.precision == '64b' else jnp.int32,
+            'bool': jnp.bool_,
         }
         self.default_do_grad = True
 
@@ -88,7 +88,7 @@ class jax_backend(object):
         Returns:
             JAX ndarray: A clipped `tensor`
         """
-        return np.clip(tensor_in, min_value, max_value)
+        return jnp.clip(tensor_in, min_value, max_value)
 
     def erf(self, tensor_in):
         """
@@ -151,7 +151,7 @@ class jax_backend(object):
         Returns:
             JAX ndarray: The tensor with repeated axes
         """
-        return np.tile(tensor_in, repeats)
+        return jnp.tile(tensor_in, repeats)
 
     def conditional(self, predicate, true_callable, false_callable):
         """
@@ -186,7 +186,7 @@ class jax_backend(object):
             raise
 
     def outer(self, tensor_in_1, tensor_in_2):
-        return np.outer(tensor_in_1, tensor_in_2)
+        return jnp.outer(tensor_in_1, tensor_in_2)
 
     def gather(self, tensor, indices):
         return tensor[indices]
@@ -195,7 +195,7 @@ class jax_backend(object):
         return tensor[mask]
 
     def isfinite(self, tensor):
-        return np.isfinite(tensor)
+        return jnp.isfinite(tensor)
 
     def astensor(self, tensor_in, dtype='float'):
         """
@@ -224,43 +224,43 @@ class jax_backend(object):
             log.error('Invalid dtype: dtype must be float, int, or bool.')
             raise
 
-        return np.asarray(tensor_in, dtype=dtype)
+        return jnp.asarray(tensor_in, dtype=dtype)
 
     def sum(self, tensor_in, axis=None):
-        return np.sum(tensor_in, axis=axis)
+        return jnp.sum(tensor_in, axis=axis)
 
     def product(self, tensor_in, axis=None):
-        return np.prod(tensor_in, axis=axis)
+        return jnp.prod(tensor_in, axis=axis)
 
     def abs(self, tensor):
-        return np.abs(tensor)
+        return jnp.abs(tensor)
 
     def ones(self, shape):
-        return np.ones(shape)
+        return jnp.ones(shape)
 
     def zeros(self, shape):
-        return np.zeros(shape)
+        return jnp.zeros(shape)
 
     def power(self, tensor_in_1, tensor_in_2):
-        return np.power(tensor_in_1, tensor_in_2)
+        return jnp.power(tensor_in_1, tensor_in_2)
 
     def sqrt(self, tensor_in):
-        return np.sqrt(tensor_in)
+        return jnp.sqrt(tensor_in)
 
     def divide(self, tensor_in_1, tensor_in_2):
-        return np.divide(tensor_in_1, tensor_in_2)
+        return jnp.divide(tensor_in_1, tensor_in_2)
 
     def log(self, tensor_in):
-        return np.log(tensor_in)
+        return jnp.log(tensor_in)
 
     def exp(self, tensor_in):
-        return np.exp(tensor_in)
+        return jnp.exp(tensor_in)
 
     def stack(self, sequence, axis=0):
-        return np.stack(sequence, axis=axis)
+        return jnp.stack(sequence, axis=axis)
 
[…]
