### 2020-10-20 a1800d0
Be more restrictive with JAX versions to reduce breaking pyhf
