### 2020-10-20 fc6941d
import shapefactor from root+xml
