### 2020-10-20 26bd9da
return None for unknown modifiers
