### 2020-10-20 55b18c3
Make sure return tensors
