### 2020-10-20 6041200
Add tests for samples
