### 2020-10-20 a2a0258
Explicilty use jax.np.asarray to avoid calling tensorlib
