### 2020-10-20 0204caa
Fixup np->jnp
