### 2020-10-20 4d64ef1
fix: Use JAX v0.2.4+ API (#1134)
* Use JAX v0.2.4 API with regards to using scipy.stats directly
* Update to jax v0.2.4+ and jaxlib v0.1.56+
   - Restrict to jax v0.2.X and jaxlib v0.1.X to ensure stability
* Add tests for Poisson and Normal sample shape
