### 2020-10-20 0ed1b51
style: Adopt new JAX style of np and jnp (#1135)
* Adopt JAX v0.2 API style change of numpy and jax.numpy imports:
   - import numpy as np
   - import jax.numpy as jnp
* c.f. https://github.com/google/jax/pull/3760
