+++ src/pyhf/tensor/jax_backend.py
@@ -2,11 +2,11 @@ from jax.config import config
 
 config.update('jax_enable_x64', True)
 
-import jax.numpy as np
+import jax.numpy as jnp
 from jax.scipy.special import gammaln
 from jax.scipy import special
 from jax.scipy.stats import norm
-import numpy as onp
+import numpy as np
 import scipy.stats as osp_stats
 import logging
 
@@ -19,9 +19,9 @@ class _BasicPoisson(object):
 
     def sample(self, sample_shape):
         # TODO: Support other dtypes
-        return np.asarray(
+        return jnp.asarray(
             osp_stats.poisson(self.rate).rvs(size=sample_shape + self.rate.shape),
-            dtype=np.float64,
+            dtype=jnp.float64,
         )
 
     def log_prob(self, value):
@@ -36,11 +36,11 @@ class _BasicNormal(object):
 
     def sample(self, sample_shape):
         # TODO: Support other dtypes
-        return np.asarray(
+        return jnp.asarray(
             osp_stats.norm(self.loc, self.scale).rvs(
                 size=sample_shape + self.loc.shape
             ),
-            dtype=np.float64,
+            dtype=jnp.float64,
         )
 
     def log_prob(self, value):
@@ -57,9 +57,9 @@ class jax_backend(object):
         self.name = 'jax'
         self.precision = kwargs.get('precision', '64b')
         self.dtypemap = {
-            'float': np.float64 if self.precision == '64b' else np.float32,
-            'int': np.int64 if self.precision == '64b' else np.int32,
-            'bool': np.bool_,
+            'float': jnp.float64 if self.precision == '64b' else jnp.float32,
+            'int': jnp.int64 if self.precision == '64b' else jnp.int32,
+            'bool': jnp.bool_,
         }
         self.default_do_grad = True
 
@@ -88,7 +88,7 @@ class jax_backend(object):
         Returns:
             JAX ndarray: A clipped `tensor`
         """
-        return np.clip(tensor_in, min_value, max_value)
+        return jnp.clip(tensor_in, min_value, max_value)
 
     def erf(self, tensor_in):
         """
@@ -151,7 +151,7 @@ class jax_backend(object):
         Returns:
             JAX ndarray: The tensor with repeated axes
         """
-        return np.tile(tensor_in, repeats)
+        return jnp.tile(tensor_in, repeats)
 
     def conditional(self, predicate, true_callable, false_callable):
         """
@@ -179,14 +179,14 @@ class jax_backend(object):
 
     def tolist(self, tensor_in):
         try:
-            return onp.asarray(tensor_in).tolist()
+            return np.asarray(tensor_in).tolist()
         except AttributeError:
             if isinstance(tensor_in, list):
                 return tensor_in
             raise
 
     def outer(self, tensor_in_1, tensor_in_2):
-        return np.outer(tensor_in_1, tensor_in_2)
+        return jnp.outer(tensor_in_1, tensor_in_2)
 
     def gather(self, tensor, indices):
         return tensor[indices]
@@ -195,7 +195,7 @@ class jax_backend(object):
         return tensor[mask]
 
     def isfinite(self, tensor):
-        return np.isfinite(tensor)
+        return jnp.isfinite(tensor)
 
     def astensor(self, tensor_in, dtype='float'):
         """
@@ -224,43 +224,43 @@ class jax_backend(object):
             log.error('Invalid dtype: dtype must be float, int, or bool.')
             raise
 
-        return np.asarray(tensor_in, dtype=dtype)
[…]
