### 2020-10-20 038836b
feat: Support ShapeFactor for import/export (#1136)
* Add ShapeFactor support to pyhf.readxml
* Add ShapeFactor support to pyhf.writexml
* Add ShapeFactor test
