### 2020-10-20 07cdf3c
Remove NumPy as dependency
