+++ src/pyhf/tensor/jax_backend.py
@@ -6,7 +6,6 @@ import jax.numpy as jnp
 from jax.scipy.special import gammaln, xlogy
 from jax.scipy import special
 from jax.scipy.stats import norm
-import numpy as np
 import scipy.stats as osp_stats
 import logging
 
