+++ src/pyhf/tensor/jax_backend.py
@@ -178,7 +178,7 @@ class jax_backend:
 
     def tolist(self, tensor_in):
         try:
-            return np.asarray(tensor_in).tolist()
+            return jnp.asarray(tensor_in).tolist()
         except AttributeError:
             if isinstance(tensor_in, list):
                 return tensor_in
