### 2020-10-20 69f7550
update set-env to use environment files
