### 2020-10-20 84381f8
chore: Migrate to new way to set environment variables in GHA (#1139)
* Migrate workflow from deprecated 'set-env' and 'add-path' to 'exportVariable' and 'addPath'
   - c.f. https://docs.github.com/en/free-pro-team@latest/actions/reference/workflow-commands-for-github-actions#environment-files
