### 2020-10-20 1e3f729
Bump version: 0.5.2 → 0.5.3
Triggered by #1139 via GitHub Actions.
