+++ setup.cfg
@@ -1,6 +1,6 @@
 [metadata]
 name = pyhf
-version = 0.5.2
+version = 0.5.3
 description = pure-Python HistFactory implementation with tensors and autodiff
 long_description = file: README.rst
 long_description_content_type = text/x-rst
