### 2020-10-20 #2734 @matthewfeickert
We just released [`v0.5.3`](https://github.com/scikit-hep/pyhf/releases/tag/v0.5.3) so I rebased the paper branch and corrected the Zenodo version number.
