### 2020-10-26 #2734 @matthewfeickert
Just wanted to give a gentle ping here to see if there was any further thoughts on having @betatim review the paper given @eloisabentivegna's request.
