### 2020-10-27 9eb8a12
Pop off the hypotest specific kwargs before passing to create_calculator
Creating objects that hold the values of the hypotest return specific
kwargs then allows for them to be efficiently retained while stripping
them from the kwargs before getting passed to create_calculator so that
there are no

TypeError: __init__() got an unexpected keyword argument

errors
