### 2020-10-27 2a98446
Use updated test stats API
