### 2020-10-27 ef6aa76
Use model suggested init_pars, par_bounds, fixed_params
