### 2020-10-27 649d6df
Fix doctest
Set qtilde=True by default for examples

Also set_backend("numpy")

The emperical distribution expected value should be cross-checked
