### 2020-10-27 a7ecbe8
Change values for qtilde_mu_bkg.samples in test_infer
