+++ tests/test_infer.py
@@ -327,16 +327,16 @@ def test_toy_calculator(tmpdir, hypotest_args):
     )
     assert qtilde_mu_bkg.samples.tolist() == pytest.approx(
         [
-            2.266462524102735,
-            1.0816608839427886,
-            2.75702209716826,
-            1.3835689306813492,
-            0.4707465263496715,
+            2.2664625238298868,
+            1.081660885985002,
+            2.757022090389057,
+            1.3835689306226868,
+            0.4707466955809423,
             0.0,
-            3.7166483729429274,
-            3.8021899113241773,
-            5.114135683200743,
-            1.3511153851299014,
+            3.7166483694865065,
+            3.8021896741039427,
+            5.114135403520493,
+            1.35111537136072,
         ],
         1e-07,
     )
