+++ src/pyhf/infer/utils.py
@@ -1,6 +1,5 @@
 """Inference for Statistical Models."""
 
-from .. import get_backend
 from .calculators import AsymptoticCalculator, ToyCalculator
 
 
@@ -37,185 +36,3 @@ def create_calculator(calctype, *args, **kwargs):
     return {'asymptotics': AsymptoticCalculator, 'toybased': ToyCalculator,}[
         calctype
     ](*args, **kwargs)
-
-
-def hypotest(
-    poi_test,
-    data,
-    pdf,
-    init_pars=None,
-    par_bounds=None,
-    fixed_params=None,
-    qtilde=False,
-    calctype="asymptotics",
-    return_tail_probs=False,
-    return_expected=False,
-    return_expected_set=False,
-    **kwargs,
-):
-    r"""
-    Compute :math:`p`-values and test statistics for a single value of the parameter of interest.
-
-    See :py:class:`~pyhf.infer.calculators.AsymptoticCalculator` and :py:class:`~pyhf.infer.calculators.ToyCalculator` on additional keyword arguments to be specified.
-
-    Example:
-        >>> import pyhf
-        >>> pyhf.set_backend("numpy")
-        >>> model = pyhf.simplemodels.hepdata_like(
-        ...     signal_data=[12.0, 11.0], bkg_data=[50.0, 52.0], bkg_uncerts=[3.0, 7.0]
-        ... )
-        >>> observations = [51, 48]
-        >>> data = pyhf.tensorlib.astensor(observations + model.config.auxdata)
-        >>> mu_test = 1.0
-        >>> CLs_obs, CLs_exp_band = pyhf.infer.hypotest(
-        ...     mu_test, data, model, return_expected_set=True
-        ... )
-        >>> CLs_obs
-        array(0.05251497)
-        >>> CLs_exp_band
-        [array(0.00260626), array(0.01382005), array(0.06445321), array(0.23525644), array(0.57303621)]
-
-    Args:
-        poi_test (Number or Tensor): The value of the parameter of interest (POI)
-        data (Number or Tensor): The data considered
-        pdf (~pyhf.pdf.Model): The statistical model adhering to the schema ``model.json``
-        init_pars (:obj:`tensor`): The initial parameter values to be used for minimization
-        par_bounds (:obj:`tensor`): The parameter value bounds to be used for minimization
-        fixed_params (:obj:`tensor`): Whether to fix the parameter to the init_pars value during minimization
-        qtilde (:obj:`bool`): When ``True`` perform the calculation using the alternative
-         test statistic, :math:`\tilde{q}_{\mu}`, as defined under the Wald
-         approximation in Equation (62) of :xref:`arXiv:1007.1727`.
-        calctype (:obj:`str`): The calculator to create. Choose either 'asymptotics' (default) or 'toybased'.
-        return_tail_probs (:obj:`bool`): Bool for returning :math:`\mathrm{CL}_{s+b}` and :math:`\mathrm{CL}_{b}`
-        return_expected (:obj:`bool`): Bool for returning :math:`\mathrm{CL}_{\mathrm{exp}}`
-        return_expected_set (:obj:`bool`): Bool for returning the :math:`(-2,-1,0,1,2)\sigma` :math:`\mathrm{CL}_{\mathrm{exp}}` --- the "Brazil band"
-
-    Returns:
-        Tuple of Floats and lists of Floats:
-
-            - :math:`\mathrm{CL}_{s}`: The modified :math:`p`-value compared to
-              the given threshold :math:`\alpha`, typically taken to be :math:`0.05`,
-              defined in :xref:`arXiv:1007.1727` as
-
-            .. math::
-
[…]
