+++ tests/test_notebooks.py
@@ -64,7 +64,7 @@ def test_impactplot(common_kwargs):
 
 @pytest.mark.slow
 def test_toys(common_kwargs):
-    pm.execute_notebook('docs/examples/notebooks/ImpactPlot.ipynb', **common_kwargs)
+    pm.execute_notebook('docs/examples/notebooks/toys.ipynb', **common_kwargs)
 
 
 def test_learn_interpolationcodes(common_kwargs):
