### 2020-10-27 bbc08ad
fix doctest
