+++ src/pyhf/tensor/tensorflow_backend.py
@@ -256,6 +256,18 @@ class tensorflow_backend(object):
     def reshape(self, tensor, newshape):
         return tf.reshape(tensor, newshape)
 
+    def ravel(self, tensor):
+        """
+        Return a flattened view of the tensor, not a copy.
+
+        Args:
+            tensor (Tensor): Tensor object
+
+        Returns:
+            `numpy.ndarray`: A flattened array.
+        """
+        return self.reshape(tensor, -1)
+
     def divide(self, tensor_in_1, tensor_in_2):
         return tf.divide(tensor_in_1, tensor_in_2)
 
