+++ src/pyhf/tensor/jax_backend.py
@@ -304,6 +304,18 @@ class jax_backend(object):
     def reshape(self, tensor, newshape):
         return jnp.reshape(tensor, newshape)
 
+    def ravel(self, tensor):
+        """
+        Return a flattened view of the tensor, not a copy.
+
+        Args:
+            tensor (Tensor): Tensor object
+
+        Returns:
+            `jax.interpreters.xla.DeviceArray`: A flattened array.
+        """
+        return np.ravel(tensor)
+
     def einsum(self, subscripts, *operands):
         """
         Evaluates the Einstein summation convention on the operands.
