### 2020-10-27 83190f1
rewrite tests slightly to account for where
