### 2020-10-27 1b40195
add test for ravel
