+++ tests/test_tensor.py
@@ -85,6 +85,22 @@ def test_tensor_where_tensor(backend):
     )
 
 
+def test_tensor_ravel(backend):
+    tb = pyhf.tensorlib
+    assert (
+        tb.tolist(
+            tb.ravel(
+                tb.astensor(
+                    [
+                        [1, 2, 3],
+                        [4, 5, 6],
+                    ]
+                )
+            )
+        )
+    ) == [1, 2, 3, 4, 5, 6]
+
+
 def test_complex_tensor_ops(backend):
     tb = pyhf.tensorlib
     assert tb.tolist(tb.outer(tb.astensor([1, 2, 3]), tb.astensor([4, 5, 6]))) == [
