### 2020-10-27 e975f3a
Add numpy example
