### 2020-10-27 ddff5fc
Add PyTorch example
