### 2020-10-27 4d738f2
Correct tensor type
