+++ src/pyhf/tensor/tensorflow_backend.py
@@ -266,7 +266,8 @@ class tensorflow_backend(object):
             >>> pyhf.set_backend("tensorflow")
             >>> tensor = pyhf.tensorlib.astensor([[1.0, 2.0, 3.0], [4.0, 5.0, 6.0]])
             >>> t_ravel = pyhf.tensorlib.ravel(tensor)
-            <tf.Tensor: shape=(6,), dtype=float32, numpy=array([1., 2., 3., 4., 5., 6.], dtype=float32)>
+            >>> print(t_ravel)
+            tf.Tensor([1. 2. 3. 4. 5. 6.], shape=(6,), dtype=float32)
 
         Args:
             tensor (Tensor): Tensor object
