### 2020-10-27 638e5c7
feat: Tensorlib ravel functionality (#1147)
* Add ravel method to tensorlib for flattened view of tensor
* Add tests for ravel
