### 2020-10-27 95fb385
Update to torch v1.7+ to use native quantile
