### 2020-10-27 3e15519
JAX now supports interpolation
