+++ src/pyhf/infer/calculators.py
@@ -128,7 +128,7 @@ class AsymptoticCalculator(object):
         init_pars=None,
         par_bounds=None,
         fixed_params=None,
-        qtilde=False,
+        qtilde=True,
     ):
         """
         Asymptotic Calculator.
