### 2020-10-28 aa4848c
Give more explicit linking for qtilde bool
Link to the test stat functions used
