### 2020-10-28 4b3e98d
Move information on the asymptotic approximations to the calculator
Mention of the approximations of the distributions of the test stats
in the later subsections of section 3 of the aysmptotics paper is moved
from the defitinion of the test stat to the asymptotic calcualtor.
