### 2020-10-28 69a5388
feat: Update model and inference defaults (#1114)
* Use the same defaults across the Python and CLI API
   - Use interpolation code4 and code4p in default modifier settings for models
   - Use qtilde as the default test stat for inference
* Link to test stat functions in docstrings
   - Migrate mention of approximation of test stat distributions to AsymptoticCalculator.distributions
