### 2020-10-28 f5a683e
Keep docstrings to 80 chars when possible
