### 2020-10-28 81c9adb
feat: Add toy calculator, empirical distribution, and toy example notebook (#790)
* Add EmpiricalDistribution and ToyCalculator classes
* Move top level infer functions under 'infer.calculators'
* Add tests for EmpiricalDistribution and ToyCalculator
* Add EmpiricalDistribution and ToyCalculator to docs
* Add example notebook on how to use the ToyCalculator

Co-authored-by: Giordon Stark <kratsg@gmail.com>
Co-authored-by: Matthew Feickert <matthew.feickert@cern.ch>
