+++ src/pyhf/infer/calculators.py
@@ -10,11 +10,25 @@ Using the calculators hypothesis tests can then be performed.
 from .mle import fixed_poi_fit
 from .. import get_backend
 from .test_statistics import qmu, qmu_tilde
+import tqdm
 
 
 def generate_asimov_data(asimov_mu, data, pdf, init_pars, par_bounds, fixed_params):
     """
     Compute Asimov Dataset (expected yields at best-fit values) for a given POI value.
+
+    Example:
+
+        >>> import pyhf
+        >>> pyhf.set_backend("numpy")
+        >>> model = pyhf.simplemodels.hepdata_like(
+        ...     signal_data=[12.0, 11.0], bkg_data=[50.0, 52.0], bkg_uncerts=[3.0, 7.0]
+        ... )
+        >>> observations = [51, 48]
+        >>> data = observations + model.config.auxdata
+        >>> mu_test = 1.0
+        >>> pyhf.infer.calculators.generate_asimov_data(mu_test, data, model, None, None, None)
+        array([ 60.61229858,  56.52802479, 270.06832542,  48.31545488])
 
     Args:
         asimov_mu (:obj:`float`): The value for the parameter of interest to be used.
@@ -66,6 +80,14 @@ class AsymptoticTestStatDistribution(object):
         """
         Compute the value of the cumulative distribution function for a given value of the test statistic.
 
+        Example:
+
+            >>> import pyhf
+            >>> pyhf.set_backend("numpy")
+            >>> bkg_dist = pyhf.infer.calculators.AsymptoticTestStatDistribution(0.0)
+            >>> bkg_dist.pvalue(0)
+            0.5
+
         Args:
             value (:obj:`float`): The test statistic value.
 
@@ -109,6 +131,15 @@ class AsymptoticTestStatDistribution(object):
         """
         Return the expected value of the test statistic.
 
+        Example:
+
+            >>> import pyhf
+            >>> pyhf.set_backend("numpy")
+            >>> bkg_dist = pyhf.infer.calculators.AsymptoticTestStatDistribution(0.0)
+            >>> n_sigma = pyhf.tensorlib.astensor([-2, -1, 0, 1, 2])
+            >>> bkg_dist.expected_value(n_sigma)
+            array([-2., -1.,  0.,  1.,  2.])
+
         Args:
             nsigma (:obj:`int` or :obj:`tensor`): The number of standard deviations.
 
@@ -142,6 +173,10 @@ class AsymptoticCalculator(object):
             qtilde (:obj:`bool`): When ``True`` use :func:`~pyhf.infer.test_statistics.qmu_tilde`
              as the test statistic.
              When ``False`` use :func:`~pyhf.infer.test_statistics.qmu`.
+            qtilde (:obj:`bool`): When ``True`` perform the calculation using the alternative test statistic,
+             :math:`\\tilde{q}`, as defined in Equation (62) of :xref:`arXiv:1007.1727`
+             (:func:`~pyhf.infer.test_statistics.qmu_tilde`).
+             When ``False`` use :func:`~pyhf.infer.test_statistics.qmu`.
 
         Returns:
             ~pyhf.infer.calculators.AsymptoticCalculator: The calculator for asymptotic quantities.
@@ -162,8 +197,24 @@ class AsymptoticCalculator(object):
         :math:`\S` 3 of :xref:`arXiv:1007.1727` under the Wald approximation,
         under the signal + background and background-only hypotheses.
 
+        Example:
+
+            >>> import pyhf
+            >>> pyhf.set_backend("numpy")
+            >>> model = pyhf.simplemodels.hepdata_like(
[…]

[…]
