### 2020-10-29 10e0b1d
get executor into toy calc
