### 2020-10-29 85934fc
handle futures subclassing
