+++ src/pyhf/infer/calculators.py
@@ -785,10 +785,6 @@ class ToyCalculator:
 
         teststat_func = utils.get_test_stat(self.test_stat)
 
-        wrapped_teststat_func = lambda sample: teststat_func(
-            poi_test, sample, self.pdf, self.init_pars, self.par_bounds, self.fixed_params
-        )
-
         tqdm_options = dict(
             total=self.ntoys,
             leave=False,
@@ -798,14 +794,45 @@ class ToyCalculator:
             unit='toy',
         )
 
+        signal_sample = tqdm.tqdm(
+            signal_sample, **tqdm_options, position=0, desc='Signal-like'
+        )
+        bkg_sample = tqdm.tqdm(
+            bkg_sample, **tqdm_options, position=1, desc='Background-like'
+        )
+
         signal_teststat = self.executor.map(
-            wrapped_teststat_func,
-            tqdm.tqdm(signal_sample, **tqdm_options, desc='Signal-like'),
+            teststat_func,
+            *zip(
+                *(
+                    (
+                        poi_test,
+                        sample,
+                        self.pdf,
+                        self.init_pars,
+                        self.par_bounds,
+                        self.fixed_params,
+                    )
+                    for sample in signal_sample
+                )
+            ),
         )
 
         bkg_teststat = self.executor.map(
-            wrapped_teststat_func,
-            tqdm.tqdm(bkg_sample, **tqdm_options, desc='Background-like'),
+            teststat_func,
+            *zip(
+                *(
+                    (
+                        poi_test,
+                        sample,
+                        self.pdf,
+                        self.init_pars,
+                        self.par_bounds,
+                        self.fixed_params,
+                    )
+                    for sample in bkg_sample
+                )
+            ),
         )
 
         s_plus_b = EmpiricalDistribution(tensorlib.astensor(list(signal_teststat)))
