### 2020-10-29 22775ca
get tqdm working for non-trivial executors
