### 2020-10-29 c4f8aed
allow for tqdm to work both ways
