+++ src/pyhf/contrib/viz/brazil.py
@@ -2,7 +2,9 @@
 import numpy as np
 
 
-def plot_results(ax, mutests, tests, test_size=0.05, alpha = 1.0, colors = ['yellow','green']):
+def plot_results(
+    ax, mutests, tests, test_size=0.05, alpha=1.0, colors=['yellow', 'green']
+):
     """Plot a series of hypothesis tests for various POI values."""
     cls_obs = np.array([test[0] for test in tests]).flatten()
     cls_exp = [np.array([test[1][i] for test in tests]).flatten() for i in range(5)]
@@ -11,7 +13,7 @@ def plot_results(ax, mutests, tests, test_size=0.05, alpha = 1.0, colors = ['yel
         ax.plot(
             mutests, cls_exp[idx], c=color, linestyle='dotted' if idx != 2 else 'dashed'
         )
-    ax.fill_between(mutests, cls_exp[0], cls_exp[-1], facecolor=colors[0], alpha = alpha)
-    ax.fill_between(mutests, cls_exp[1], cls_exp[-2], facecolor=colors[1], alpha = alpha)
+    ax.fill_between(mutests, cls_exp[0], cls_exp[-1], facecolor=colors[0], alpha=alpha)
+    ax.fill_between(mutests, cls_exp[1], cls_exp[-2], facecolor=colors[1], alpha=alpha)
     ax.plot(mutests, [test_size] * len(mutests), c='red')
     ax.set_ylim(0, 1)
