+++ src/pyhf/infer/calculators.py
@@ -551,7 +551,7 @@ class ToyCalculator:
                     poi_test,
                     sample,
                     self.pdf,
-                    signal_pars,
+                    self.init_pars,
                     self.par_bounds,
                     self.fixed_params,
                 )
@@ -564,7 +564,7 @@ class ToyCalculator:
                     poi_test,
                     sample,
                     self.pdf,
-                    bkg_pars,
+                    self.init_pars,
                     self.par_bounds,
                     self.fixed_params,
                 )
