### 2020-11-04 a8b578b
add ROOT cross-check
