### 2020-11-04 #2734 @eloisabentivegna
> Hi, I will be happy to do this. Please provide me with the instructions and timeline of what is expected out of me.

Suchita, thanks for agreeing to help! As you might have gathered if you've looked around this repository a little, a JOSS review is centred upon downloading and installing the software, and inspecting the repository and submitted paper for key elements. You can find much more details at https://joss.readthedocs.io/en/latest/reviewer_guidelines.html, but I am also happy to elaborate.

Since this is much more interactive than traditional reviews, we encourage you to start as soon as possible. We would expect things to wrap up within six weeks, at the latest.
