### 2020-11-05 9c37fcb
pyupgrade to 3.6+
