+++ src/pyhf/writexml.py
@@ -300,4 +300,7 @@ def writexml(spec, specdir, data_rootdir, resultprefix):
     for measurement in spec['measurements']:
         combination.append(build_measurement(measurement, dict(mixin.modifiers)))
     indent(combination)
-    return "<!DOCTYPE Combination  SYSTEM 'HistFactorySchema.dtd'>\n\n".encode() + ET.tostring(combination, encoding='utf-8')
+    return (
+        "<!DOCTYPE Combination  SYSTEM 'HistFactorySchema.dtd'>\n\n".encode()
+        + ET.tostring(combination, encoding='utf-8')
+    )
