+++ validation/xmlimport_input2/makedata.py
@@ -17,9 +17,7 @@ for cname, channel_def in source_data['channels'].iteritems():
 
     for hist, data in bindata.iteritems():
         print(f'{cname}_{hist}')
-        h = ROOT.TH1F(
-            f'{cname}_{hist}', f'{cname}_{hist}', *binning
-        )
+        h = ROOT.TH1F(f'{cname}_{hist}', f'{cname}_{hist}', *binning)
         hists += [h]
         for i, v in enumerate(data):
             h.SetBinContent(i + 1, v)
