### 2020-11-05 a36ecec
Add pyupgrade to pre-commit hooks
