+++ .pre-commit-config.yaml
@@ -4,3 +4,7 @@ repos:
     hooks:
     - id: black
       language_version: python3
+-   repo: https://github.com/asottile/pyupgrade
+    rev: v2.7.3
+    hooks:
+    -   id: pyupgrade
