+++ src/pyhf/writexml.py
@@ -300,7 +300,6 @@ def writexml(spec, specdir, data_rootdir, resultprefix):
     for measurement in spec['measurements']:
         combination.append(build_measurement(measurement, dict(mixin.modifiers)))
     indent(combination)
-    return (
-        b"<!DOCTYPE Combination  SYSTEM 'HistFactorySchema.dtd'>\n\n"
-        + ET.tostring(combination, encoding='utf-8')
+    return b"<!DOCTYPE Combination  SYSTEM 'HistFactorySchema.dtd'>\n\n" + ET.tostring(
+        combination, encoding='utf-8'
     )
