+++ src/pyhf/utils.py
@@ -39,7 +39,7 @@ def validate(spec, schema_name, version=None):
     schema = load_schema(schema_name, version=version)
     try:
         resolver = jsonschema.RefResolver(
-            base_uri="file://{pkg_resources.resource_filename(__name__, 'schemas/'):s}",
+            base_uri=f"file://{pkg_resources.resource_filename(__name__, 'schemas/'):s}",
             referrer=schema_name,
             store=SCHEMA_CACHE,
         )
@@ -53,7 +53,7 @@ def validate(spec, schema_name, version=None):
 
 def options_from_eqdelimstring(opts):
     document = '\n'.join(
-        "{opt.split('=', 1)[0]}: {opt.split('=', 1)[1]}" for opt in opts
+        f"{opt.split('=', 1)[0]}: {opt.split('=', 1)[1]}" for opt in opts
     )
     return yaml.full_load(document)
 
