+++ src/pyhf/workspace.py
@@ -347,9 +347,7 @@ class Workspace(_ChannelSummaryMixin, dict):
         elif self.measurement_names:
             if measurement_name is not None:
                 if measurement_name not in self.measurement_names:
-                    log.debug(
-                        f"measurements defined: {self.measurement_names:s}"
-                    )
+                    log.debug(f"measurements defined: {self.measurement_names:s}")
                     raise exceptions.InvalidMeasurement(
                         f'no measurement by name \'{measurement_name:s}\' was found in the workspace, pick from one of the valid ones above'
                     )
