### 2020-11-05 67ba7fa
Use f-strings in tests
