+++ tests/test_examples.py
@@ -2,6 +2,6 @@ import shlex
 
 
 def test_2bin_1channel(tmpdir, script_runner):
-    command = 'pyhf inspect {:s}'.format('docs/examples/json/2-bin_1-channel.json')
+    command = f"pyhf inspect {'docs/examples/json/2-bin_1-channel.json':s}"
     ret = script_runner.run(*shlex.split(command))
     assert ret.success
