+++ tests/test_scripts.py
@@ -27,9 +27,7 @@ def test_version(script_runner):
 # see test_import.py for the same (detailed) test
 def test_import_prepHistFactory(tmpdir, script_runner):
     temp = tmpdir.join("parsed_output.json")
-    command = 'pyhf xml2json validation/xmlimport_input/config/example.xml --basedir validation/xmlimport_input/ --output-file {:s} --hide-progress'.format(
-        temp.strpath
-    )
+    command = f'pyhf xml2json validation/xmlimport_input/config/example.xml --basedir validation/xmlimport_input/ --output-file {temp.strpath:s} --hide-progress'
     ret = script_runner.run(*shlex.split(command))
     assert ret.success
     assert ret.stdout == ''
@@ -42,9 +40,7 @@ def test_import_prepHistFactory(tmpdir, script_runner):
 
 def test_import_prepHistFactory_withProgress(tmpdir, script_runner):
     temp = tmpdir.join("parsed_output.json")
-    command = 'pyhf xml2json validation/xmlimport_input/config/example.xml --basedir validation/xmlimport_input/ --output-file {:s}'.format(
-        temp.strpath
-    )
+    command = f'pyhf xml2json validation/xmlimport_input/config/example.xml --basedir validation/xmlimport_input/ --output-file {temp.strpath:s}'
     ret = script_runner.run(*shlex.split(command))
     assert ret.success
     assert ret.stdout == ''
@@ -63,9 +59,7 @@ def test_import_prepHistFactory_stdout(tmpdir, script_runner):
 
 def test_import_prepHistFactory_and_fit(tmpdir, script_runner):
     temp = tmpdir.join("parsed_output.json")
-    command = "pyhf xml2json validation/xmlimport_input/config/example.xml --basedir validation/xmlimport_input/ --output-file {:s}".format(
-        temp.strpath
-    )
+    command = f"pyhf xml2json validation/xmlimport_input/config/example.xml --basedir validation/xmlimport_input/ --output-file {temp.strpath:s}"
     ret = script_runner.run(*shlex.split(command))
 
     command = f"pyhf fit {temp.strpath:s}"
@@ -83,9 +77,7 @@ def test_import_prepHistFactory_and_fit(tmpdir, script_runner):
         "LogNormExample",
         "ConstExample",
     ]:
-        command = "pyhf fit {:s} --value --measurement {:s}".format(
-            temp.strpath, measurement
-        )
+        command = f"pyhf fit {temp.strpath:s} --value --measurement {measurement:s}"
         ret = script_runner.run(*shlex.split(command))
 
         assert ret.success
@@ -106,9 +98,7 @@ def test_import_prepHistFactory_and_fit(tmpdir, script_runner):
 
 def test_import_prepHistFactory_and_cls(tmpdir, script_runner):
     temp = tmpdir.join("parsed_output.json")
-    command = 'pyhf xml2json validation/xmlimport_input/config/example.xml --basedir validation/xmlimport_input/ --output-file {:s}'.format(
-        temp.strpath
-    )
+    command = f'pyhf xml2json validation/xmlimport_input/config/example.xml --basedir validation/xmlimport_input/ --output-file {temp.strpath:s}'
     ret = script_runner.run(*shlex.split(command))
 
     command = f'pyhf cls {temp.strpath:s}'
@@ -148,9 +138,7 @@ def test_import_prepHistFactory_and_cls(tmpdir, script_runner):
 @pytest.mark.parametrize("backend", ["numpy", "tensorflow", "pytorch", "jax"])
 def test_fit_backend_option(tmpdir, script_runner, backend):
     temp = tmpdir.join("parsed_output.json")
[…]
