+++ tests/conftest.py
@@ -108,18 +108,14 @@ def backend(request):
 
     if skip_backend and (param_id in only_backends):
         raise ValueError(
-            "Must specify skip_{param} or only_{param} but not both!".format(
-                param=param_id
-            )
+            f"Must specify skip_{param_id} or only_{param_id} but not both!"
         )
 
     if skip_backend:
         pytest.skip(f"skipping {func_name} as specified")
     elif only_backends and param_id not in only_backends:
         pytest.skip(
-            "skipping {func} as specified to only look at: {backends}".format(
-                func=func_name, backends=', '.join(only_backends)
-            )
+            f"skipping {func_name} as specified to only look at: {', '.join(only_backends)}"
         )
 
     if fail_backend:
