+++ tests/test_backend_consistency.py
@@ -137,17 +137,13 @@ def test_hypotest_qmu_tilde(
         assert (numpy_ratio_delta_unity < tolerance['numpy']).all()
     except AssertionError:
         print(
-            'Ratio to NumPy+SciPy exceeded tolerance of {}: {}'.format(
-                tolerance['numpy'], numpy_ratio_delta_unity.tolist()
-            )
+            f"Ratio to NumPy+SciPy exceeded tolerance of {tolerance['numpy']}: {numpy_ratio_delta_unity.tolist()}"
         )
         assert False
     try:
         assert (tensors_ratio_delta_unity < tolerance['tensors']).all()
     except AssertionError:
         print(
-            'Ratio between tensor backends exceeded tolerance of {}: {}'.format(
-                tolerance['tensors'], tensors_ratio_delta_unity.tolist()
-            )
+            f"Ratio between tensor backends exceeded tolerance of {tolerance['tensors']}: {tensors_ratio_delta_unity.tolist()}"
         )
         assert False
