### 2020-11-05 b9d6026
Use f-strings in validation
