+++ validation/manualonoff_roofit/onoff.py
@@ -95,4 +95,4 @@ c.SaveAs('scan.pdf')
 
 print(f'observed: {result.UpperLimit()}')
 for i in [-2, -1, 0, 1, 2]:
-    print('expected {}: {}'.format(i, result.GetExpectedUpperLimit(i)))
+    print(f'expected {i}: {result.GetExpectedUpperLimit(i)}')
