### 2020-11-05 93c7bba
Fix bad operand for adding strings error
fixes

TypeError: bad operand type for unary +: 'str'
