### 2020-11-05 a3b34a5
chore: Run pyupgrade on codebase (3.6+) (#1164)
* Run pyupgrade --py36-plus on codebase
* Add pyupgrade to pre-commit hooks
* Add pre-commit.ci status badge to README
* Use f-strings over format() as much as possible
* Thanks to Anthony Sottile (@asottile) for pyupgrade and pre-commit.ci

Co-authored-by: Matthew Feickert <matthew.feickert@cern.ch>
