+++ src/pyhf/cli/patchset.py
@@ -54,7 +54,7 @@ def extract(patchset, name, output_file, with_metadata):
     if output_file:
         with open(output_file, 'w+') as out_file:
             json.dump(result, out_file, indent=4, sort_keys=True)
-        log.debug("Written to {0:s}".format(output_file))
+        log.debug(f"Written to {output_file:s}")
     else:
         click.echo(json.dumps(result, indent=4, sort_keys=True))
 
@@ -93,7 +93,7 @@ def apply(background_only, patchset, name, output_file):
     if output_file:
         with open(output_file, 'w+') as out_file:
             json.dump(patched_ws, out_file, indent=4, sort_keys=True)
-        log.debug("Written to {0:s}".format(output_file))
+        log.debug(f"Written to {output_file:s}")
     else:
         click.echo(json.dumps(patched_ws, indent=4, sort_keys=True))
 
