+++ tests/test_init.py
@@ -35,7 +35,7 @@ def test_missing_backends(isolate_modules, param):
     CACHE_BACKEND, sys.modules[backend_name] = sys.modules[backend_name], None
     sys.modules.setdefault(f'pyhf.tensor.{import_name}', None)
     CACHE_MODULE, sys.modules[f'pyhf.tensor.{module_name}'] = (
-        sys.modules['pyhf.tensor.{}'.format(module_name)],
+        sys.modules[f'pyhf.tensor.{module_name}'],
         None,
     )
 
@@ -73,9 +73,9 @@ def test_missing_optimizer(isolate_modules, param):
 
     # hide
     CACHE_BACKEND, sys.modules[backend_name] = sys.modules[backend_name], None
-    sys.modules.setdefault('pyhf.optimize.{}'.format(import_name), None)
-    CACHE_MODULE, sys.modules['pyhf.optimize.{}'.format(import_name)] = (
-        sys.modules['pyhf.optimize.{}'.format(import_name)],
+    sys.modules.setdefault(f'pyhf.optimize.{import_name}', None)
+    CACHE_MODULE, sys.modules[f'pyhf.optimize.{import_name}'] = (
+        sys.modules[f'pyhf.optimize.{import_name}'],
         None,
     )
     try:
@@ -88,7 +88,7 @@ def test_missing_optimizer(isolate_modules, param):
 
     # put back
     CACHE_BACKEND, sys.modules[backend_name] = None, CACHE_BACKEND
-    CACHE_MODULE, sys.modules['pyhf.optimize.{}'.format(import_name)] = (
+    CACHE_MODULE, sys.modules[f'pyhf.optimize.{import_name}'] = (
         None,
         CACHE_MODULE,
     )
