+++ .pre-commit-config.yaml
@@ -7,4 +7,9 @@ repos:
 -   repo: https://github.com/asottile/pyupgrade
     rev: v2.7.3
     hooks:
-    -   id: pyupgrade
+    - id: pyupgrade
+-   repo: https://github.com/nbQA-dev/nbQA
+    rev: 0.4.0
+    hooks:
+    - id: nbqa-black
+    - id: nbqa-pyupgrade
