### 2020-11-06 17dc90d
Add py36-plus option to pyupgrade hook
