+++ .pre-commit-config.yaml
@@ -8,6 +8,7 @@ repos:
     rev: v2.7.3
     hooks:
     - id: pyupgrade
+      args: ["--py36-plus"]
 -   repo: https://github.com/nbQA-dev/nbQA
     rev: 0.4.0
     hooks:
