### 2020-11-06 e4011ff
feat: Add nbQA to pre-commit hooks (#1165)
* Add nbqa to pre-commit hooks
   - Apply nbqa-black and nbqa-pyupgrade
* Add nbqa config options to pyproject.toml
* Run nbqa over all notebooks
* Add --py36-plus option as arg to pyupgrade pre-commit hook
   - Amends PR #1164
