+++ pyproject.toml
@@ -62,3 +62,13 @@ markers = [
     "skip_pytorch64",
     "skip_tensorflow",
 ]
+
+[tool.nbqa.config]
+black = "pyproject.toml"
+
+[tool.nbqa.mutate]
+black = 1
+pyupgrade = 1
+
+[tool.nbqa.addopts]
+pyupgrade = ["--py36-plus"]
