### 2020-11-06 e44b265
Restrict to uproot4 v0.1.X
