+++ src/pyhf/writexml.py
@@ -5,8 +5,10 @@ import shutil
 import pkg_resources
 import xml.etree.cElementTree as ET
 import numpy as np
-import uproot4 as uproot
-from uproot4.behaviors.TH1 import TH1
+import uproot as uproot3
+from uproot_methods.classes import TH1
+
+# TODO: Move to uproot4 when ROOT file writing is supported
 
 from .mixins import _ChannelSummaryMixin
 
@@ -32,9 +34,9 @@ def _make_hist_name(channel, sample, modifier='', prefix='hist', suffix=''):
 
 
 def _export_root_histogram(histname, data):
-    hist = TH1((np.asarray(data), np.arange(len(data) + 1)))
+    hist = TH1.from_numpy((np.asarray(data), np.arange(len(data) + 1)))
     hist._fName = histname
-    # FIXME: uproot crashes for some reason, figure out why later
+    # FIXME: uproot3 crashes for some reason, figure out why later
     # if histname in _ROOT_DATA_FILE:
     #    raise KeyError(f'Duplicate key {histname} being written.')
     _ROOT_DATA_FILE[histname] = hist
@@ -270,7 +272,7 @@ def writexml(spec, specdir, data_rootdir, resultprefix):
         "Combination", OutputFilePrefix=str(Path(specdir).joinpath(resultprefix))
     )
 
-    with uproot.recreate(
+    with uproot3.recreate(
         str(Path(data_rootdir).joinpath('data.root'))
     ) as _ROOT_DATA_FILE:
         for channelspec in spec['channels']:
