### 2020-11-07 #2823 @whedon
[REVIEW]: pyhf: pure-Python implementation of HistFactory statistical models
**Submitting author:** @matthewfeickert (<a href="http://orcid.org/0000-0003-4124-7862">Matthew Feickert</a>)
**Repository:** <a href="https://github.com/scikit-hep/pyhf" target ="_blank">https://github.com/scikit-hep/pyhf</a>
**Version:** v0.5.4
