### 2020-11-08 45f128d
Pin additional dependencies in pre-commit
