+++ .pre-commit-config.yaml
@@ -13,4 +13,6 @@ repos:
     rev: 0.4.0
     hooks:
     - id: nbqa-black
+      additional_dependencies: [black==20.8b1]
     - id: nbqa-pyupgrade
+      additional_dependencies: [pyupgrade==2.7.3]
