+++ docs/contributors.rst
@@ -18,3 +18,4 @@ Contributors include:
 - Nikolai Hartmann
 - Alexander Held
 - Karthikeyan Singaravelan
+- Marco Gorelli
