+++ .pre-commit-config.yaml
@@ -3,7 +3,6 @@ repos:
     rev: 20.8b1
     hooks:
     - id: black
-      language_version: python3
 -   repo: https://github.com/asottile/pyupgrade
     rev: v2.7.3
     hooks:
