### 2020-11-09 28fdfe9
fix: Pin additional dependencies in pre-commit hooks (#1167)
* Set dependency versions of black and pyupgrade for nbQA to use
* Remove language_version from Black pre-commit hook as already specified
