+++ .pre-commit-config.yaml
@@ -3,7 +3,6 @@ repos:
     rev: 20.8b1
     hooks:
     - id: black
-      language_version: python3
 -   repo: https://github.com/asottile/pyupgrade
     rev: v2.7.3
     hooks:
@@ -13,4 +12,6 @@ repos:
     rev: 0.4.0
     hooks:
     - id: nbqa-black
+      additional_dependencies: [black==20.8b1]
     - id: nbqa-pyupgrade
+      additional_dependencies: [pyupgrade==2.7.3]
