### 2020-11-10 b8a0082
more explicit var names
