### 2020-11-10 904f652
Add docstrings for emperical distributions methods
