### 2020-11-10 99edeae
Make floats instead of tesnors for consistency
Check on this
