+++ docs/api.rst
@@ -121,6 +121,7 @@ Inference
 
 .. autosummary::
    :toctree: _generated/
+   :template: modifierclass.rst
 
    test_statistics.q0
    test_statistics.qmu
