### 2020-11-10 5ef5352
Simplify and use more explicit naming
Change name from CLs_exp to CLs_exp_band to make clear not a singular
value
