### 2020-11-10 03c444a
Add pyhf version to hypo test demo
