### 2020-11-10 1cbda00
add notes on test stat definition
