### 2020-11-10 3b93c12
Restructure to allow multiple runs
