### 2020-11-12 0548901
Add stub of docstring
