### 2020-11-12 3b6d0fe
Add label settings
