### 2020-11-12 66dab0b
Add altair as extra
