### 2020-11-13 53c423b
Add docstrings examle
