### 2020-11-13 7c8fe43
feat: Add default axis labels to contrib.viz.brazil.plot_results (#1176)
* Add default xlabel and ylabel to pyhf.contrib.viz.brazil.plot_results axis
   - Update baseline image for tests
* Add docstring and example
