+++ .pre-commit-config.yaml
@@ -4,12 +4,12 @@ repos:
     hooks:
     - id: black
 -   repo: https://github.com/asottile/pyupgrade
-    rev: v2.7.3
+    rev: v2.7.4
     hooks:
     - id: pyupgrade
       args: ["--py36-plus"]
 -   repo: https://github.com/nbQA-dev/nbQA
-    rev: 0.4.0
+    rev: 0.4.1
     hooks:
     - id: nbqa-black
       additional_dependencies: [black==20.8b1]
