### 2020-11-16 821dfe5
chore: Update pyupgrade and nbQA in pre-commit (#1178)
* [pre-commit.ci] pre-commit autoupdate
   - Update pyupgrade to v2.7.4
   - Update nbQA to v0.4.1 and pyupgrade dependency
