+++ .pre-commit-config.yaml
@@ -4,14 +4,14 @@ repos:
     hooks:
     - id: black
 -   repo: https://github.com/asottile/pyupgrade
-    rev: v2.7.3
+    rev: v2.7.4
     hooks:
     - id: pyupgrade
       args: ["--py36-plus"]
 -   repo: https://github.com/nbQA-dev/nbQA
-    rev: 0.4.0
+    rev: 0.4.1
     hooks:
     - id: nbqa-black
       additional_dependencies: [black==20.8b1]
     - id: nbqa-pyupgrade
-      additional_dependencies: [pyupgrade==2.7.3]
+      additional_dependencies: [pyupgrade==2.7.4]
