### 2020-11-18 97e2647
Add test that will fail if intergers not correctly cast to floats
