### 2020-11-18 2356b07
Move test JSON to dir and use datadir fixture
