### 2020-11-18 d83771e
fix: Ensure float-like values used in division in writexml (#1179)
* Ensure float-like values are used in division by using float dtype
* Add test for float-like values using integer spec data
