+++ docs/bib/tutorials.bib
@@ -1,6 +1,17 @@
 % NB: entries with same author-title-year are not picked up:
 %     https://github.com/mcmtroffaes/sphinxcontrib-bibtex/issues/117
 
+@unpublished{GStark20200925,
+  title  = {{ATLAS Exotics + SUSY Workshop 2020 pyhf Tutorial}},
+  author = {Giordon Stark},
+  year   = {2020},
+  month  = {September},
+  day    = {25},
+  note   = {ATLAS Exotics + SUSY Workshop 2020},
+  organization = {CERN},
+  url    = {https://pyhf.github.io/tutorial-ATLAS-SUSY-Exotics-2020/introduction.html},
+}
+
 @unpublished{Feickert20200716,
   title  = {{pyhf: Accelerating analyses and preserving likelihoods}},
   author = {Matthew Feickert},
