### 2020-11-18 f6bc6f3
use strategy 0 for user-provided gradient
