### 2020-11-18 9f24b5a
Update src/pyhf/optimize/opt_minuit.py
Co-authored-by: Matthew Feickert <matthew.feickert@cern.ch>
