### 2020-11-18 4a162ef
add the other tests to check configuration
