### 2020-11-18 9a2da57
better configuration
