### 2020-11-19 985aafa
feat: Set minuit strategy automatically for gradient/non-gradient mode (#1183)
* Set MINUIT strategy to 0/1 automatically if using gradients/no-gradient
* Allow for MINUIT strategy to be configurable
* Add test for strategy selection
