### 2020-11-19 0bc205e
Add use citation in CHEP 2019 proceedings
