+++ docs/bib/use_citations.bib
@@ -1,3 +1,17 @@
+% 2020-11-16
+@article{Feickert:2020chep2019,
+    author = "Feickert, Matthew and Heinrich, Lukas and Stark, Giordon",
+    title = "{Likelihood preservation and statistical reproduction of searches for new physics}",
+    primaryClass = "hep-ex",
+    doi = "10.1051/epjconf/202024506017",
+    journal = "EPJ Web Conf.",
+    volume = "245",
+    number = "06017",
+    pages = "7",
+    year = "2020"
+}
+
+% 2020-09-03
 @article{Alguero:2020grj,
     author = "Alguero, Gaël and Kraml, Sabine and Waltenberger, Wolfgang",
     title = "{A SModelS interface for pyhf likelihoods}",
