+++ docs/bib/use_citations.bib
@@ -7,7 +7,6 @@
     journal = "EPJ Web Conf.",
     volume = "245",
     number = "06017",
-    pages = "7",
     year = "2020"
 }
 
