### 2020-11-19 ab93698
add test for setting tolerance for each mle.fit
