### 2020-11-19 ef95a59
add test for setting tolerance
