### 2020-11-19 2010b6d
docs: Add CHEP 2019 proceedings use citation (#1185)
* Add use citation for the published CHEP 2019 conference proceedings
   - c.f. https://doi.org/10.1051/epjconf/202024506017
