### 2020-11-19 07fc449
feat: Configurable optimizer tolerance for termination (#1184)
* Add configurable optimizer tolerance for minimization termination
   - NB: SciPy defaults to None, iminuit defaults to 0.1
* Add tests to ensure the API for SciPy/iminuit
