### 2020-11-19 a373668
Guard to catch at CLI first use
