### 2020-11-19 33205ce
fix: Isolate pyhf contrib commands from rest of CLI (#1186)
* Move import pyhf.contrib.utils inside the contrib CLI functions
   - Guard the CLI API from missing extras without triggering warnings from using other modules CLI API
* Update the modules to reimport in tests
