### 2020-11-19 8497db9
apply iminuit PRs 518 and 519
