+++ src/pyhf/modifiers/shapefactor.py
@@ -14,7 +14,6 @@ class shapefactor:
         return {
             'paramset_type': unconstrained,
             'n_parameters': len(sample_data),
-            'modifier': cls.__name__,
             'is_constrained': cls.is_constrained,
             'is_shared': True,
             'inits': (1.0,) * len(sample_data),
