### 2020-11-19 ee246bf
refactor: Drop 'modifier' from parameter set requirements (#1189)
* Drop the modifier from required_parset()
   - Parameter set requirements do not need to know modifier
