+++ src/pyhf/modifiers/lumi.py
@@ -14,7 +14,6 @@ class lumi:
         return {
             'paramset_type': constrained_by_normal,
             'n_parameters': 1,
-            'modifier': cls.__name__,
             'is_constrained': cls.is_constrained,
             'is_shared': True,
             'op_code': cls.op_code,
