### 2020-11-23 a33badd
chore: Update nbQA in pre-commit (#1190)
* [pre-commit.ci] pre-commit autoupdate
   - Update nbQA to v0.5.0
