### 2020-11-23 fc891ee
Use uproot3 PyPI namespace
