+++ src/pyhf/pdf.py
@@ -731,7 +731,7 @@ class Model:
             ):  # force to be not scalar, should we changed with #522
                 return tensorlib.reshape(result, (1,))
             return result
-        except:
+        except:  # noqa: E722
             log.error(
                 f'eval failed for data {tensorlib.tolist(data)} pars: {tensorlib.tolist(pars)}'
             )
