### 2020-11-24 6757700
Fix E713
src/pyhf/readxml.py:43:8: E713 test for membership should be 'not in'
