### 2020-11-24 e7d4661
Add flake8 to pre-commit hooks
