+++ .pre-commit-config.yaml
@@ -3,6 +3,10 @@ repos:
     rev: 20.8b1
     hooks:
     - id: black
+-   repo: https://gitlab.com/pycqa/flake8
+    rev: 3.8.4
+    hooks:
+    - id: flake8
 -   repo: https://github.com/asottile/pyupgrade
     rev: v2.7.4
     hooks:
