### 2020-11-24 1e64b92
Move flake8 config to setup.cfg
