+++ setup.cfg
@@ -60,6 +60,6 @@ all-files = 1
 warning-is-error = 1
 
 [flake8]
-extend-ignore = E203, E266, E402, E501
+extend-ignore = E203, E402, E501
 max-line-length = 88
 select = B,C,E,F,W,T4,B9
