### 2020-11-24 3d59f5d
Update to specify uproot3
