+++ src/pyhf/writexml.py
@@ -5,7 +5,7 @@ import shutil
 import pkg_resources
 import xml.etree.cElementTree as ET
 import numpy as np
-import uproot
+import uproot3 as uproot
 from uproot_methods.classes import TH1
 
 from .mixins import _ChannelSummaryMixin
