+++ tests/test_import.py
@@ -1,7 +1,7 @@
 import pyhf
 import pyhf.readxml
 import numpy as np
-import uproot
+import uproot3 as uproot
 from pathlib import Path
 import pytest
 import xml.etree.cElementTree as ET
