### 2020-11-24 e9d25ac
Remove uproot from test extra
