+++ setup.cfg
@@ -60,6 +60,9 @@ all-files = 1
 warning-is-error = 1
 
 [flake8]
+# E203: whitespace before ':'
+# E402: module level import not at top of file
+# E501: line too long
 extend-ignore = E203, E402, E501
 max-line-length = 88
 select = B,C,E,F,W,T4,B9
