### 2020-11-24 28413ec
Select all and enable count stats
