+++ setup.cfg
@@ -65,4 +65,5 @@ warning-is-error = 1
 # E501: line too long
 extend-ignore = E203, E402, E501
 max-line-length = 88
-select = B,C,E,F,W,T4,B9
+count = True
+statistics = True
