### 2020-11-24 fbe0252
chore: Use flake8 for linting (#1191)
* Switch from pyflakes to flake8 for linting in pre-commit hooks and CI
   - Use flake8 in 'lint' extra
* Add flake8 config options to setup.cfg (as opposed to .flake8)
* Remove code that existed only to placate pyflakes
