+++ setup.cfg
@@ -58,3 +58,12 @@ source-dir = docs
 build-dir = docs/_build
 all-files = 1
 warning-is-error = 1
+
+[flake8]
+# E203: whitespace before ':'
+# E402: module level import not at top of file
+# E501: line too long
+extend-ignore = E203, E402, E501
+max-line-length = 88
+count = True
+statistics = True
