### 2020-11-24 6160064
Remove noqa E722 from pdf.py
