### 2020-11-24 c799943
Use repr for better message
Thanks to Marco Gorelli for the tip!
