### 2020-11-24 b96fdc9
Use exc_info=True to get info on exception
