### 2020-11-25 2a08a0b
verify that minuit parameter uncertainties match expectation
