### 2020-11-25 f7e9a85
verify that NLL matches expectation
