### 2020-11-25 0c8c0c4
test: Verify NLL values and Minuit parameter uncertainties (#1197)
* Add a test for the NLL value returned by minimization
* Explicitly specify the fixed_vals kwarg for function calls passing in fixed parameter values
