+++ tests/test_optim.py
@@ -351,7 +351,7 @@ def test_optim(backend, source, spec, mu):
         pdf,
         init_pars,
         par_bounds,
-        [(pdf.config.poi_index, mu)],
+        fixed_vals=[(pdf.config.poi_index, mu)],
     )
     assert pyhf.tensorlib.tolist(result)
 
@@ -375,11 +375,12 @@ def test_optim_with_value(backend, source, spec, mu):
         pdf,
         init_pars,
         par_bounds,
-        [(pdf.config.poi_index, mu)],
+        fixed_vals=[(pdf.config.poi_index, mu)],
         return_fitted_val=True,
     )
     assert pyhf.tensorlib.tolist(result)
     assert pyhf.tensorlib.shape(fitted_val) == ()
+    assert pytest.approx(17.52954975, rel=1e-5) == fitted_val
 
 
 @pytest.mark.parametrize('mu', [1.0], ids=['mu=1'])
@@ -402,11 +403,11 @@ def test_optim_uncerts(backend, source, spec, mu):
         pdf,
         init_pars,
         par_bounds,
-        [(pdf.config.poi_index, mu)],
+        fixed_vals=[(pdf.config.poi_index, mu)],
         return_uncertainties=True,
     )
     assert result.shape[1] == 2
-    assert pyhf.tensorlib.tolist(result)
+    assert pytest.approx([0.0, 0.26418431]) == pyhf.tensorlib.tolist(result[:, 1])
 
 
 @pytest.mark.parametrize(
