+++ .pre-commit-config.yaml
@@ -13,7 +13,7 @@ repos:
     - id: pyupgrade
       args: ["--py36-plus"]
 -   repo: https://github.com/nbQA-dev/nbQA
-    rev: 0.5.0
+    rev: 0.5.1
     hooks:
     - id: nbqa-black
       additional_dependencies: [black==20.8b1]
