### 2020-11-30 aa17119
chore: Specify exceptions to remove inline ignores of E722 (#1194)
* Specify exception for all exception cases
* Remove all inline noqa: E722
   - E722: Do not use bare except, specify exception instead
