### 2020-11-30 4b25692
fix typos in docstrings
