### 2020-11-30 755fca9
docs: Correct spelling of continuous (#1201)
* Fix typo of the word "continuous" in a few docstrings and a comment
   - continous -> continuous
