### 2020-11-30 e7cdbc8
Simply try to pass all to Minuit init
