### 2020-11-30 ecdb945
Try to pass init pars in correctly to *args
