### 2020-11-30 1b6b0e7
Try without mocking from_array_func
