### 2020-11-30 f3d14e8
Remove from_array_func mocking
