### 2020-11-30 2d1b8ef
feat: Use iminuit v1.5.X API (#1202)
* Require iminuit compatible with v1.5.3 in 'minuit' extra
* Adopt iminuit v1.5.3+ API
   - ncalls -> ncalls_total
   - ngrads -> ngrads_total
* Remove mocking of iminuit.Minuit().from_array_func from tests
   - c.f. https://github.com/scikit-hep/iminuit/pull/464
