### 2020-12-03 fe520e9
Exclude older python version of MacOS in testing
