+++ .github/workflows/ci.yml
@@ -14,10 +14,12 @@ jobs:
     runs-on: ${{ matrix.os }}
     strategy:
       matrix:
-        os: [ubuntu-latest, macos-latest]
+        os: [ubuntu-latest, macos-latest, windows-latest]
         python-version: [3.6, 3.7, 3.8]
         exclude:
           - os: macos-latest
+            python-version: [3.6, 3.7]
+          - os: windows-latest
             python-version: [3.6, 3.7]
 
     steps:
