### 2020-12-03 344ab68
Revert "Try to add Windows testing"
This reverts commit ae2d071e4570985a5d4e4e4662caee3ae51542af.
