+++ .github/workflows/ci.yml
@@ -18,7 +18,9 @@ jobs:
         python-version: [3.6, 3.7, 3.8]
         exclude:
           - os: macos-latest
-            python-version: [3.6, 3.7]
+            python-version: 3.6
+          - os: macos-latest
+            python-version: 3.7
 
     steps:
     - uses: actions/checkout@v2
