### 2020-12-03 62466c3
ci: Test only latest Python runtime on MacOS (#1203)
* Exclude Python 3.6 and Python 3.7 from the testing matrix for macos-latest
   - Speed up CI significantly while still testing the latest Python runtime on macos-latest
* Add testing of Python 3.6 to current release tests
* c.f. https://docs.github.com/en/free-pro-team@latest/actions/reference/workflow-syntax-for-github-actions#example-excluding-configurations-from-a-matrix
