+++ .github/workflows/ci.yml
@@ -16,6 +16,11 @@ jobs:
       matrix:
         os: [ubuntu-latest, macos-latest]
         python-version: [3.6, 3.7, 3.8]
+        exclude:
+          - os: macos-latest
+            python-version: 3.6
+          - os: macos-latest
+            python-version: 3.7
 
     steps:
     - uses: actions/checkout@v2
