### 2020-12-07 d64e819
DeviceArray to _DeviceArray
