+++ .pre-commit-config.yaml
@@ -13,7 +13,7 @@ repos:
     - id: pyupgrade
       args: ["--py36-plus"]
 -   repo: https://github.com/nbQA-dev/nbQA
-    rev: 0.5.1
+    rev: 0.5.4
     hooks:
     - id: nbqa-black
       additional_dependencies: [black==20.8b1]
