### 2020-12-09 #2823 @eloisabentivegna
@suchitakulkarni, @bradkav, were you able to start the review process, e.g. downloading `pyhf`? Could you post an update?
