### 2020-12-09 e3c4cad
use exc_info and remove unneeded error captures
