+++ src/pyhf/tensor/jax_backend.py
@@ -221,7 +221,9 @@ class jax_backend:
         try:
             dtype = self.dtypemap[dtype]
         except KeyError:
-            log.error('Invalid dtype: dtype must be float, int, or bool.')
+            log.error(
+                'Invalid dtype: dtype must be float, int, or bool.', exc_info=True
+            )
             raise
 
         return jnp.asarray(tensor_in, dtype=dtype)
