### 2020-12-09 6045839
move citation into pyhf data
