### 2020-12-09 95cb957
add cli for citation
