+++ tests/test_utils.py
@@ -64,3 +64,11 @@ def test_remove_prefix():
     assert pyhf.utils.remove_prefix('abcDEF123', 'abc') == 'DEF123'
     assert pyhf.utils.remove_prefix('abcDEF123', 'Abc') == 'abcDEF123'
     assert pyhf.utils.remove_prefix('abcDEF123', '123') == 'abcDEF123'
+
+
+@pytest.mark.parametrize('oneline', [False, True])
+def test_citation(oneline):
+    citation = pyhf.utils.citation(oneline)
+    assert citation
+    if oneline:
+        assert '\n' not in citation
