### 2020-12-09 f584d53
add doctstrings
