### 2020-12-09 fd0bb76
feat: Add utility to get software citation and CLI interface (#1210)
* Add citation utility to pyhf
* Add CLI functionality to print out pyhf citation
