+++ src/pyhf/data/citation.bib
@@ -1,0 +1,7 @@
+@software{pyhf,
+  author = "{Heinrich, Lukas and Feickert, Matthew and Stark, Giordon}",
+  title = "{pyhf: v0.5.3}",
+  version = {0.5.3},
+  doi = {10.5281/zenodo.1169739},
+  url = {https://github.com/scikit-hep/pyhf},
+}
