+++ .bumpversion.cfg
@@ -9,7 +9,7 @@ tag = True
 
 [bumpversion:file:README.rst]
 
-[bumpversion:file:docs/bib/preferred.bib]
+[bumpversion:file:src/pyhf/data/citation.bib]
 
 [bumpversion:file:.zenodo.json]
 
