### 2020-12-09 8a06262
fix: Use uproot3 namespace for patch release v0.5.4
* Install uproot3 in xmlio extra
* import uproot3 as uproot
