+++ src/pyhf/writexml.py
@@ -5,8 +5,8 @@ import shutil
 import pkg_resources
 import xml.etree.cElementTree as ET
 import numpy as np
-import uproot
-from uproot_methods.classes import TH1
+import uproot3 as uproot
+from uproot3_methods.classes import TH1
 
 from .mixins import _ChannelSummaryMixin
 
