### 2020-12-09 956ed6e
fix: Update JAX DeviceArray type in docstrings for doctest
* Update JAX DeviceArray type in docstrings to jax.interpreters.xla._DeviceArray
   - Needed to pass doctest
