+++ src/pyhf/version.py
@@ -2,4 +2,4 @@
 
 # Use semantic versioning (https://semver.org/)
 # The version number is controlled through bumpversion.cfg
-__version__ = '0.5.3'
+__version__ = '0.5.4'
