+++ README.rst
@@ -281,8 +281,8 @@ BibTeX entry for citation of ``pyhf`` is
 
    @software{pyhf,
      author = "{Heinrich, Lukas and Feickert, Matthew and Stark, Giordon}",
-     title = "{pyhf: v0.5.3}",
-     version = {0.5.3},
+     title = "{pyhf: v0.5.4}",
+     version = {0.5.4},
      doi = {10.5281/zenodo.1169739},
      url = {https://github.com/scikit-hep/pyhf},
    }
