### 2020-12-10 6f1ac62
chore: Use exc_info in log.error, re-raise last exceptions (#1209)
* Use exc_info in log.error to show error + message
* Use raise without arguments to re-raise last raised exception
