+++ src/pyhf/cli/rootio.py
@@ -38,7 +38,8 @@ def xml2json(entrypoint_xml, basedir, output_file, track_progress):
     except ImportError:
         log.error(
             "xml2json requires uproot, please install pyhf using the "
-            "xmlio extra: python -m pip install pyhf[xmlio]"
+            "xmlio extra: python -m pip install pyhf[xmlio]",
+            exc_info=True,
         )
     from .. import readxml
 
@@ -67,7 +68,8 @@ def json2xml(workspace, output_dir, specroot, dataroot, resultprefix, patch):
     except ImportError:
         log.error(
             "json2xml requires uproot, please install pyhf using the "
-            "xmlio extra: python -m pip install pyhf[xmlio]"
+            "xmlio extra: python -m pip install pyhf[xmlio]",
+            exc_info=True,
         )
     from .. import writexml
 
