### 2020-12-10 #2823 @matthewfeickert
@eloisabentivegna @suchitakulkarni @bradkav, in the time since we first submitted `pyhf` to JOSS pre-review we've also release patch release [`v0.5.3`](https://github.com/scikit-hep/pyhf/releases/tag/v0.5.3) and are about to release `v0.5.4`. Do you want us to rebase the branch that has the JOSS paper in with these number updated, or would you prefer everything remain strictly frozen at [`v0.5.2`](https://github.com/scikit-hep/pyhf/releases/tag/v0.5.2)?
