### 2020-12-10 924b86c
docs: Link to conda-forge feedstock and link banner to repo (#1217)
- Link conda-forge badge to pyhf-feedstock repo
   - c.f. https://github.com/conda-forge/pyhf-feedstock
- Make docs banner link to repo page and not fork page
   - Change banner text from 'Fork me' to 'View me'
