+++ docs/index.rst
@@ -23,7 +23,7 @@
 
 .. raw:: html
 
-   <a class="github-fork-ribbon right-top fixed" href="https://github.com/scikit-hep/pyhf/fork" data-ribbon="Fork me on GitHub" title="Fork me on GitHub">Fork me on GitHub</a>
+   <a class="github-fork-ribbon right-top fixed" href="https://github.com/scikit-hep/pyhf/" data-ribbon="View me on GitHub" title="View me on GitHub">View me on GitHub</a>
 
 .. include:: ../README.rst
 
