### 2020-12-11 #2823 @matthewfeickert
> given we are in the early stages of the review process, I think a rebase would not be too disruptive. I leave it up to the authors: if you prefer to publish the latest version with JOSS, please make sure the paper branch is up to date. Once the reviews are further under way, it may be best to keep the code frozen.

Sounds good! I just rebased so I'll trigger wheadon now.
