### 2020-12-11 057221d
Set values through OO API now
