### 2020-12-11 99c28b1
np_covariance -> covariance
* from CHANGELOG: Minuit.np_covariance is now obsolete and was removed; see Minuit.covariance
