+++ src/pyhf/optimize/opt_minuit.py
@@ -143,7 +143,7 @@ class minuit_optimizer(OptimizerMixin):
         if minimizer.valid:
             # Extra call to hesse() after migrad() is always needed for good error estimates. If you pass a user-provided gradient to MINUIT, convergence is faster.
             minimizer.hesse()
-            hess_inv = minimizer.np_covariance()
+            hess_inv = minimizer.covariance()
 
         unc = None
         if return_uncertainties:
