### 2020-12-11 bd802d3
np_values -> values
* Minuit.np_values was removed; use Minuit.values instead or np.array(m.values)
