### 2020-12-11 3a247e9
np_errors -> errors
* Minuit.np_errors was removed; use Minuit.errors instead or np.array(m.errors)
