+++ tests/test_optim.py
@@ -496,22 +496,34 @@ def test_init_pars_sync_fixed_values_scipy(mocker):
 def test_init_pars_sync_fixed_values_minuit(mocker):
     opt = pyhf.optimize.minuit_optimizer()
 
-    # patch all we need
-    from pyhf.optimize import opt_minuit
+    def f(x, y, z):
+        # Need a function of three parameters for below
+        return (x - 2) ** 2 + (y - 3) ** 2 + (z - 4) ** 2
 
-    minimizer = mocker.patch.object(opt_minuit, 'iminuit')
-    opt._get_minimizer(None, [9, 9, 9], [(0, 10)] * 3, fixed_vals=[(0, 1)])
-    assert minimizer.Minuit.from_array_func.call_args[1]['start'] == [1, 9, 9]
-    assert minimizer.Minuit.from_array_func.call_args[1]['fix'] == [True, False, False]
+    # @kratsg, What is the need for mocker here?
+    # # patch all we need
+    # from pyhf.optimize import opt_minuit
+    #
+    # minimizer = mocker.patch.object(opt_minuit, 'iminuit')
+    minuit = opt._get_minimizer(f, [9, 9, 9], [(0, 10)] * 3, fixed_vals=[(0, 1)])
+    # assert minimizer.Minuit.from_array_func.call_args[1]['start'] == [1, 9, 9]
+    # assert minimizer.Minuit.from_array_func.call_args[1]['fix'] == [True, False, False]
+    assert list(minuit.values) == [1, 9, 9]
+    assert list(minuit.fixed) == [True, False, False]
 
 
 def test_step_sizes_fixed_parameters_minuit(mocker):
     opt = pyhf.optimize.minuit_optimizer()
 
-    # patch all we need
-    from pyhf.optimize import opt_minuit
+    def f(x, y, z):
+        # Need a function of three parameters for below
+        return (x - 2) ** 2 + (y - 3) ** 2 + (z - 4) ** 2
 
-    minimizer = mocker.patch.object(opt_minuit, 'iminuit')
-    opt._get_minimizer(None, [9, 9, 9], [(0, 10)] * 3, fixed_vals=[(0, 1)])
-    assert minimizer.Minuit.from_array_func.call_args[1]['fix'] == [True, False, False]
-    assert minimizer.Minuit.from_array_func.call_args[1]['error'] == [0.0, 0.01, 0.01]
+    # @kratsg, What is the need for mocker here?
+    # # patch all we need
+    # from pyhf.optimize import opt_minuit
+    #
+    # minimizer = mocker.patch.object(opt_minuit, 'iminuit')
+    minuit = opt._get_minimizer(f, [9, 9, 9], [(0, 10)] * 3, fixed_vals=[(0, 1)])
+    assert list(minuit.fixed) == [True, False, False]
+    assert list(minuit.errors) == [0.0, 0.01, 0.01]
