### 2020-12-11 6baee20
Remove old code
