### 2020-12-13 940449e
Update bib to v0.5.4
