+++ docs/JOSS/paper.bib
@@ -58,8 +58,8 @@
 
 @software{pyhf_zenodo,
   author = "{Heinrich, Lukas and Feickert, Matthew and Stark, Giordon}",
-  title = "{pyhf: v0.5.3}",
-  version = {0.5.3},
+  title = "{pyhf: v0.5.4}",
+  version = {0.5.4},
   doi = {10.5281/zenodo.1169739},
   url = {https://github.com/scikit-hep/pyhf},
   year = {2020}
