### 2020-12-13 #2823 @matthewfeickert
Rebased to get `v0.5.4` ([detailed here](https://github.com/scikit-hep/pyhf/releases/tag/v0.5.4)) which is the same as `v0.5.3` but with some safeguarding of optional dependencies changes.
