+++ .github/workflows/dependencies-head.yml
@@ -57,7 +57,7 @@ jobs:
       run: |
         python -m pytest -r sx --ignore tests/benchmarks/ --ignore tests/contrib --ignore tests/test_notebooks.py
 
-  uproot:
+  uproot3:
 
     runs-on: ${{ matrix.os }}
     strategy:
@@ -75,7 +75,7 @@ jobs:
       run: |
         python -m pip install --upgrade pip setuptools wheel
         python -m pip install --ignore-installed --upgrade -q --no-cache-dir -e .[test]
-        python -m pip install --upgrade --no-cache-dir git+git://github.com/scikit-hep/uproot.git
+        python -m pip install --upgrade --no-cache-dir git+git://github.com/scikit-hep/uproot3.git
         python -m pip list
     - name: Test with pytest
       run: |
