### 2020-12-13 524fc96
Restrict sphinxcontrib-bibtex
