### 2020-12-13 4f3bd17
Build sphinx second time
c.f. https://sphinxcontrib-bibtex.readthedocs.io/en/latest/usage.html#running-sphinx
