+++ .github/workflows/docs.yml
@@ -45,6 +45,7 @@ jobs:
         python -m doctest README.rst
         python setup.py build_sphinx
         touch docs/_build/html/.nojekyll
+        python setup.py build_sphinx
     - name: Check schemas are copied over
       run: |
         # is a directory
