+++ .github/workflows/docs.yml
@@ -46,6 +46,7 @@ jobs:
         cd docs
         make html
         make html # Second time for sphinxcontrib-bibtex
+        cd ..
         touch docs/_build/html/.nojekyll
     - name: Check schemas are copied over
       run: |
