### 2020-12-13 a35161e
fix: Ensure pytest.approx only compares floats (#1220)
* Compare floats to floats and not tensors in pytest.approx
* Add pytest and uproot3 to HEAD of dependencies workflow
* Restrict sphinxcontrib-bibtex to v1.X
   - c.f. PR #1221
