### 2020-12-13 7b0eb57
Use Python 3.8 for RTD builds
