### 2020-12-14 6b0a931
chore: Update to nbQA v0.5.5 in pre-commit (#1222)
* [pre-commit.ci] pre-commit autoupdate
   - Update nbQA to v0.5.5
