### 2020-12-15 scikit-hep/pyhf#1228 @matthewfeickert
Make a scikit-hep-bot
I want a scikit-hep-bot that is as cool as @pytestbot or @whedon! :)

c.f. https://github.com/pytest-dev/pytest/issues/8148 and https://github.com/openjournals/joss-reviews/issues/2823 for examples of what this kind of bot could do.
