### 2020-12-15 scikit-hep/pyhf#1228 @matthewfeickert
cc @eduardo-rodrigues and @henryiii as well to get more feedback.
