### 2020-12-15 scikit-hep/pyhf#1228 @eduardo-rodrigues
The JOSS one is clever ... scary :D. What would be the main goal for us? Across packages?

That's something totally different but it reminds me somehow of https://github.com/scikit-hep/scikit-hep.github.io/issues/42 on a news section for the website.
