### 2020-12-16 scikit-hep/pyhf#1228 @matthewfeickert
> What would be the main goal for us? Across packages?

I think that's unclear at the moment. I _really_ like the way that pytestbot is able to setup a release through a request through a comment. I think that speaks a bit to my enjoyment in the release workflow that `pyhf` has at the moment which is automated through applying tags to PRs before we merge them. I think that we'd probably have to play around with the underlying technologies and systems to understand what things are capable of though and also if it would make sense to try to have a Scikit-HEP wide bot or if we would have them for individual projects.
