### 2020-12-17 scikit-hep/pyhf#1228 @henryiii
I really like simplicity and using `setuptools_scm` - I really don't see how a dedicated UI/tool, like clicking "make new release" in the GitHub UI or in the `gh` CLI is possibly harder than using bots and PRs. I probably should try it eventually, but boost-histogram, particle, etc. are all very easy to release.
