### 2020-12-17 scikit-hep/pyhf#1228 @kratsg
We rely on `setuptools_scm` for doing the release. It's just kind of awkward when we want to release via PRs instead of pushing directly to master. We can't get around it, but we've reorganized our workflow such that a specifically-tagged PR will trigger a `bump2version` and push to master for us.
