### 2020-12-18 scikit-hep/pyhf#1228 @henryiii
Sorry, didn't mean to sound so negative there, and some of my issues would be solved by having a centralized bot. I work with a lot of repos, so steps that don't happen often and are hard to test like releases need to be simple and clean. I can't set up a complex system in all Scikit-HEP repos, and then maintain them.

`setuptools_scm` normally means you use your version control system as your versioning system. No versions in the repo files. At all. You hard code the version in setup.cfg and a bunch of other places, and therefore use bump2version, not `setuptools_scm` even if you use it for something else.

In boost-histogram, the release process is literally the GitHub release UI and nothing more. I have a bot that bugs me on each PR that doesn't update the changelog, so that stays up to date. The version numbers does not exist in the repo anywhere. Every commit gets a unique version based on the tag, the distance from the tag, and dirty vs. clean. The only manual step is I copy the release notes into the GitHub release, though I could setup actions to modify the release after the fact, I think.
