### 2020-12-18 scikit-hep/pyhf#1228 @matthewfeickert
> Sorry, didn't mean to sound so negative there, and some of my issues would be solved by having a centralized bot.

No you're fine. I don't think anyone viewed it as being negative, just stating your views and workflows.

> I can't set up a complex system in all Scikit-HEP repos, and then maintain them.

Yeah, this is why for doing releases for `pylhe` I just have it [setup to trigger releases to PyPI through cutting releases on the GitHub UI](https://github.com/scikit-hep/pylhe/blob/5a78db5608c081c9e8b5f2915887135e417f42cd/.github/workflows/publish-package.yml#L9-L10). Very simple and it works.

> `setuptools_scm` normally means you use your version control system as your versioning system. No versions in the repo files. At all. You hard code the version in setup.cfg and a bunch of other places, and therefore use bump2version, not `setuptools_scm` even if you use it for something else.

Yeah. In my mind the main way we use `setuptools_scm` is to ensure that the test release we push out to TestPyPI with each merge of a PR gets its own unique (dev) release number.

> In boost-histogram, the release process is literally the GitHub release UI and nothing more. ... The version numbers does not exist in the repo anywhere. Every commit gets a unique version based on the tag, the distance from the tag, and dirty vs. clean.

Ah, I didn't realize this! This is actually quite interesting, I'll have to look through it.

> I have a bot that bugs me on each PR that doesn't update the changelog, so that stays up to date.

We've been including the commit summary in the release notes thanks to the bot we have for a while, but I've only recently started curating them to be informative for readers. We really should start a CHANGELOG as well so that we have that on our docs website to make it easier for people to pinpout where, when, and why things got changed.
