### 2020-12-19 fd7930c
feat: Test statistic specified by string instead of using qtilde kwarg (#1231)
* Deprecate qtilde in favor of test_stat keyword argument
* Test statistic now specified as a string, instead of relying just on qtilde=True/False
