+++ src/pyhf/exceptions/__init__.py
@@ -91,6 +91,12 @@ class InvalidInterpCode(Exception):
     """
 
 
+class InvalidTestStatistic(Exception):
+    """
+    InvalidTestStatistic is raised when an invalid/unimplemented test statistic is requested.
+    """
+
+
 class ImportBackendError(Exception):
     """
     MissingLibraries is raised when something is imported by sustained an import error due to missing additional, non-default libraries.
