### 2020-12-19 #2823 @bradkav
Review:

pyhf is an incredibly useful tool, which is already an important part of the HEP analysis ecosystem. Installation is straightforward and there are plenty of examples. Before I can recommend publication in JOSS, I just have a few minor comments and suggestions, aimed mostly at making the code more accessible to newcomers.

- [ ]  **Documentation:** Some brief examples/documentation are provided in the README. However, it appears that the most complete documentation is at https://scikit-hep.org/pyhf/, so it would be useful to add a link to this more prominently in the README. This would also lead new users more quickly to the many examples at https://scikit-hep.org/pyhf/examples.html.

- [ ] **Examples:** The command-line example in the README - `cat << EOF  | tee likelihood.json | pyhf cls` - doesn't seem to work. Is there a missing `likelihood.json` file?

- [ ] **Tests:** The examples in the README give the correct output. However, it appears that there are also [automated tests in the repo](https://github.com/scikit-hep/pyhf/tree/master/tests). Is there a straightforward way for the user to run these test and verify that everyone is working fine?

- [ ] **Paper:** The [citation syntax](https://rmarkdown.rstudio.com/authoring_bibliographies_and_citations.html#citation_syntax) specifies that citations should be surrounded with square brackets and separated by semi-colons. Please could you update the paper?

- [ ] **Paper:** Typo in the paper - under "Future Work", " aims to provide support limit setting" should presumably be "aims to provide support for limit setting".
