### 2020-12-19 #2823 @kratsg
Hi @bradkav  - thanks for reviewing. For the example in the README, you should use the full block of code to the last `EOF`, that is (see below). Admittedly, this is probably not so obvious for users not familiar with all the features on the command line, so if you have suggestions on how to rephrase this, let us know!

For "Tests" - the straightforward way is through `pytest` - and just running `pytest` without any options should run all tests, if the user has checked out the development install of `pyhf`. How would you like us to document the "straightforward" way or is it enough to mention that we use `pytest` for all of our tests?

For the rest, we'll work on implementing the changes/fixes. Thanks again.

------------

```
cat << EOF  | tee likelihood.json | pyhf cls
{
    "channels": [
        { "name": "singlechannel",
          "samples": [
            { "name": "signal",
              "data": [12.0, 11.0],
              "modifiers": [ { "name": "mu", "type": "normfactor", "data": null} ]
            },
            { "name": "background",
              "data": [50.0, 52.0],
              "modifiers": [ {"name": "uncorr_bkguncrt", "type": "shapesys", "data": [3.0, 7.0]} ]
            }
          ]
        }
    ],
    "observations": [
        { "name": "singlechannel", "data": [51.0, 48.0] }
    ],
    "measurements": [
        { "name": "Measurement", "config": {"poi": "mu", "parameters": []} }
    ],
    "version": "1.0.0"
}
EOF
```

as in this screenshot:
