### 2020-12-20 #2823 @matthewfeickert
Thanks for the review @bradkav! We really appreciate it. :slightly_smiling_face: I've had whedon update the draft with the corrections (c.f. above comment).

> * [ ]   **Documentation:** Some brief examples/documentation are provided in the README. However, it appears that the most complete documentation is at https://scikit-hep.org/pyhf/, so it would be useful to add a link to this more prominently in the README. This would also lead new users more quickly to the many examples at https://scikit-hep.org/pyhf/examples.html.

There is currently a badge at the top of the README that links to the documentation:

seen also in this screenshot

![README_screen_shot](https://user-images.githubusercontent.com/5142394/102720432-eeefad80-42b9-11eb-9613-d38813e69bf2.png)

The link is also found additionally in the "About" column on the GitHub page

![about_column](https://user-images.githubusercontent.com/5142394/102720444-02027d80-42ba-11eb-9c50-9390ffbff4a8.png)

We also encourage acquiring `pyhf` from our officially supported distributions on [PyPI](https://pypi.org/project/pyhf/) and [Conda-forge](https://github.com/conda-forge/pyhf-feedstock), where the links to the documentation are clearly visible on the pages thanks to the nice page design by the orgs.

[![PyPI_page](https://user-images.githubusercontent.com/5142394/102720449-09c22200-42ba-11eb-87e5-8c1728bbacb4.png)](https://pypi.org/project/pyhf/)

[![conda-forge_page](https://user-images.githubusercontent.com/5142394/102720453-1181c680-42ba-11eb-896c-c8ed94a303a6.png)](https://github.com/conda-forge/pyhf-feedstock)

Let us know if you have additional suggestions.

> * [ ]  **Examples:** The command-line example in the README - `cat << EOF  | tee likelihood.json | pyhf cls` - doesn't seem to work. Is there a missing `likelihood.json` file?

As @kratsg has already mentioned, this example is using [heredoc formatting](https://en.wikipedia.org/wiki/Here_document#Unix_shells), which is quite common in our target audience of users.

> * [ ]  **Tests:** The examples in the README give the correct output. However, it appears that there are also [automated tests in the repo](https://github.com/scikit-hep/pyhf/tree/master/tests). Is there a straightforward way for the user to run these test and verify that everyone is working fine?

We do not distribute the test suite with the PyPI and Conda-forge distributions, as testing the software is the responsibility of the development team not the users (though we do acknowledge that libraries like NumPy and SciPy do distribute the tests in their releases). Distributing the tests would also require a large number of additional dependencies to run the full test suite (as the test suite compares execution across _all_ optional externals and backends), which users should not be required to have for runtime use. Execution of the full test suite on every commit [using GitHub Actions workflows](https://github.com/scikit-hep/pyhf/tree/master/.github/workflows) is [publicly viewable on GitHub](https://github.com/scikit-hep/pyhf/actions), and the status is viewable through the live updating badges at the top of the README for the test suite, build of the documentation, and latest deployment to PyPI:

[![CI/CD](https://github.com/scikit-hep/pyhf/workflows/CI/CD/badge.svg?branch=master)](https://github.com/scikit-hep/pyhf/actions?query=workflow%3ACI%2FCD+branch%3Amaster) [![Docs](https://github.com/scikit-hep/pyhf/workflows/Docs/badge.svg?branch=master)](https://github.com/scikit-hep/pyhf/actions?query=workflow%3ADocs+branch%3Amaster) [![publish distributions](https://github.com/scikit-hep/pyhf/workflows/publish%20distributions/badge.svg?branch=master)](https://github.com/scikit-hep/pyhf/actions?query=workflow%3A%22publish+distributions%22+branch%3Amaster)

If a user wants to run the full test suite there are details in [the developer docs page](https://scikit-hep.org/pyhf/development.html) as well as in the [`CONTRIBUTING.md`](https://github.com/scikit-hep/pyhf/blob/fd7930cce36cbc3a2d0ee1828f060d7382129579/CONTRIBUTING.md).

> * [ ]  **Paper:** The [citation syntax](https://rmarkdown.rstudio.com/authoring_bibliographies_and_citations.html#citation_syntax) specifies that citations should be surrounded with square brackets and separated by semi-colons. Please could you update the paper?

Sure, I've changed them here. In pre-review @kthyng [had suggested we remove the `[]` syntax](https://github.com/openjournals/joss-reviews/issues/2734#issuecomment-706239615). Both citations styles [differ from what is common in high energy physics](https://github.com/openjournals/joss-reviews/issues/2734#issuecomment-706256648), so we're happy to adopt either of them. We had also missed the semicolons for separation &mdash; sorry about that &mdash; so that is fixed now too.

> * [ ]  **Paper:** Typo in the paper - under "Future Work", " aims to provide support limit setting" should presumably be "aims to provide support for limit setting".

Doh! Thanks for catching that. Fixed!
