### 2020-12-22 #2823 @bradkav
Thanks for the quick reply @matthewfeickert and @kratsg!

Concerning the documentation, I now appreciate that there are quite a few different links, so it was perhaps me not looking in the right places. I'm happy with this now.

Concerning the tests, I now see the section on tests in [`CONTRIBUTING.md`](https://github.com/scikit-hep/pyhf/blob/fd7930cce36cbc3a2d0ee1828f060d7382129579/CONTRIBUTING.md), so I'm happy with that. I agree that it's not necessary to distribute the test suite (especially with the live test status badges in the repo).

The final thing is the command line example. Thanks @kratsg for clear explanation. It wasn't a familiar format to me and I tried a few different combinations of things, but never the correct one. Doh. I realise that this format is common among the target audience, but in the interest of clarity, can I suggest splitting up the input and output in the example? E.g.:

```bash
$ cat << EOF  | tee likelihood.json | pyhf cls
{
    "channels": [
        { "name": "singlechannel",
          "samples": [
            { "name": "signal",
              "data": [12.0, 11.0],
              "modifiers": [ { "name": "mu", "type": "normfactor", "data": null} ]
            },
            { "name": "background",
              "data": [50.0, 52.0],
              "modifiers": [ {"name": "uncorr_bkguncrt", "type": "shapesys", "data": [3.0, 7.0]} ]
            }
          ]
        }
    ],
    "observations": [
        { "name": "singlechannel", "data": [51.0, 48.0] }
    ],
    "measurements": [
        { "name": "Measurement", "config": {"poi": "mu", "parameters": []} }
    ],
    "version": "1.0.0"
}
EOF
```
which should produce the following JSON output:
```bash
{
   "CLs_exp": [
      0.0026062609501074576,
      0.01382005356161206,
      0.06445320535890459,
      0.23525643861460702,
      0.573036205919389
   ],
   "CLs_obs": 0.05251497423736956
}
```
