### 2021-01-07 #2823 @matthewfeickert
@eloisabentivegna @bradkav Just popping in here to note that I'll get some follow up to @bradkav's last review &mdash; thanks for it too! &mdash; (and anything that @suchitakulkarni has) soon. The `pyhf` dev team all took the winter holidays off entirely and we're now trying to fight through 2 weeks of email, but we'll reply soon.
