### 2021-01-12 #2823 @suchitakulkarni
P.S. If invited again, I will make sure to complete the review today.

On Tue, 12 Jan 2021 at 14:33, Suchita Kulkarni <suchita.kulkarni@gmail.com>
wrote:

> Hi,
>
> I was trying to complete my review today, however I can no longer access
> any of the documents. Was the invitation revoked? I am sorry to have been
> completely silent, I was trying to move to another city (my job is in Graz
> and am stuck in Vienna). I had to confirm and reconfirm everything multiple
> times due to ever changing lockdown rules in Austria.
>
> I can still complete review if it is feasible for you.
>
> I am very sorry once again!
>
> best
> Suchita.
>
> On Thu, 7 Jan 2021 at 22:42, Matthew Feickert <notifications@github.com>
> wrote:
>
>> @eloisabentivegna <https://github.com/eloisabentivegna> @bradkav
>> <https://github.com/bradkav> Just popping in here to note that I'll get
>> some follow up to @bradkav <https://github.com/bradkav>'s last review —
>> thanks for it too! — (and anything that @suchitakulkarni
>> <https://github.com/suchitakulkarni> has) soon. The pyhf dev team all
>> took the winter holidays off entirely and we're now trying to fight through
>> 2 weeks of email, but we'll reply soon.
>>
>> —
>> You are receiving this because you were mentioned.
>> Reply to this email directly, view it on GitHub
>> <https://github.com/openjournals/joss-reviews/issues/2823#issuecomment-756403161>,
>> or unsubscribe
>> <https://github.com/notifications/unsubscribe-auth/ABLGPKOA22FEZDWYK5AESCDSYYTFRANCNFSM4TN2E64A>
>> .
>>
>
>
> --
> (Suchita Kulkarni)
>

--
(Suchita Kulkarni)
