### 2021-01-19 #2823 @suchitakulkarni
Hi,

Thanks for this incredibly useful and fast package. I have a few questions/comments.

1) In the example as given at the repository, should there be a plt.show() or plt.savefig() command in the end? The plot didn't show up at least for me otherwise. It may also be very nice to provide that example as a python file such that user can run it and get 'instant satisfaction' of seeing an output :)
2) I could not run pytest on my laptop. It fails with an ugly trace and the error seems to be
'E   TypeError: function() got an unexpected keyword argument 'jit_compile'
3) Does the code has to have tensorflow/pytorch installed? If so, should it be mentioned as necessary packages? I didn't have them installed but it seems some of the tests use it?
