### 2021-01-19 #2823 @bradkav
Thanks @matthewfeickert! I'm happy with all the changes.

So @eloisabentivegna I'm now happy to recommend this for publication in JOSS :fireworks:
