### 2021-01-19 #2823 @matthewfeickert
Thanks for the review @suchitakulkarni! We've updated the `README` to reflect your feedback.

> 1. In the example as given at the repository, should there be a plt.show() or plt.savefig() command in the end? The plot didn't show up at least for me otherwise. It may also be very nice to provide that example as a python file such that user can run it and get 'instant satisfaction' of seeing an output :)

It is uncommon to give explicit figure save instructions in most libraries that build ontop of existing visualization tools, however, given your suggestions we've adopted `matplotlib`'s approach in their documentation of including an explicit `fig.show()` in the two visualization examples in the `README`. I'm not sure what you mean by "provide that example as a Python file" though. How do you envision this file being distributed to a user? The distributions of `pyhf` on PyPI and Conda-forge are distributions of the runtime, not of the entire repository. The examples in the `README` are runnable code that can be copied into any file or REPL and are shortened versions of [longer examples taken from the documentation](https://scikit-hep.org/pyhf/examples.html).

> 2. I could not run pytest on my laptop. It fails with an ugly trace and the error seems to be
>    'E   TypeError: function() got an unexpected keyword argument 'jit_compile'

As we mentioned in [our reply to @bradkav's first round of comments](https://github.com/openjournals/joss-reviews/issues/2823#issuecomment-748639982), similar to how we do not distribute the tests with distributions of `pyhf` on PyPI or Conda-forge, to run the test suite developers should install all necessary dependencies as detailed in [the developer docs page](https://scikit-hep.org/pyhf/development.html) as well as in the [`CONTRIBUTING.md`](https://github.com/scikit-hep/pyhf/blob/fd7930cce36cbc3a2d0ee1828f060d7382129579/CONTRIBUTING.md).

> 3. Does the code has to have tensorflow/pytorch installed? If so, should it be mentioned as necessary packages? I didn't have them installed but it seems some of the tests use it?

No, they are not required for user runtime as they are optional backends as [mentioned in the installation documentation](https://scikit-hep.org/pyhf/installation.html). They are of course required for the tests.
