### 2021-01-27 #2823 @matthewfeickert
> @bradkav, @suchitakulkarni, many thanks for your help and comments. Based on your recommendation, I will start the prepublication phase for this submission.

Yes, many thanks to @bradkav and @suchitakulkarni not only for their review but also for helping to make `pyhf` better! We truly do appreciate it and your time and effort. :bow:

> ```
> Reference check summary (note 'MISSING' DOIs are suggestions that need verification):
>
> MISSING DOIs
>
> - 10.2172/1614734 may be a valid DOI for title: Reinterpretation of LHC Results for New Physics: Status and Recommendations after Run 2
> ```

Ah it seems that [Reinterpretation of LHC Results for New Physics: Status and Recommendations after Run 2](https://inspirehep.net/literature/1785921) now has an updated DOI of [0.21468/SciPostPhys.9.2.022](https://doi.org/10.21468/SciPostPhys.9.2.022). I'll update this and trigger `whedon` again.
