### 2021-01-27 #2823 @eloisabentivegna
The proof reads fine to me!

@matthewfeickert, @kratsg, we'll also need a tagged release and a code archive for your submission (see https://joss.readthedocs.io/en/latest/submitting.html#the-review-process). I see that a Zenodo archive already exists, but I believe this does not match the current version. The archive title should also match the submission title.
