### 2021-01-27 #2823 @matthewfeickert
> we'll also need a tagged release and a code archive for your submission (see https://joss.readthedocs.io/en/latest/submitting.html#the-review-process). I see that a Zenodo archive already exists, but I believe this does not match the current version. The archive title should also match the submission title.

Sounds good :+1: . I'll get on that in the morning my time (US) and then update here as soon as that's taken care of.
