### 2021-01-27 #2823 @matthewfeickert
@eloisabentivegna We have merged the branch and have cut a [release on GitHub with tag `JOSS-paper`](https://github.com/scikit-hep/pyhf/releases/tag/JOSS-paper) which now exists on Zenodo as DOI [![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.4472736.svg)](https://doi.org/10.5281/zenodo.4472736).

The Zenodo title is that of the paper, "pyhf: pure-Python implementation of HistFactory statistical models", and I have set the version to be JOSS-paper instead of v0.5.4 to help distinguish it from the [`v0.5.4` release Zenodo archive](https://zenodo.org/record/4318533).

Please let me know if you have any requested changes or if there are more standard ways that people prepare JOSS papers on Zenodo.
