### 2021-01-28 #2823 @matthewfeickert
> Do you want to change the reference in the main text to point to this new archive?

@eloisabentivegna Does this show up anywhere other than this Issue? If so, and it shows up in the paper then if we could have version `v0.5.4`, [which is the release that the code reviewed corresponds to](https://github.com/openjournals/joss-reviews/issues/2823#issuecomment-743950999), that would be better. If it is just for bookkeeping then it is fine to keep it as `JOSS-paper`.
