### 2021-01-28 #2823 @eloisabentivegna
@matthewfeickert, thanks for the explanation, I didn't realize the top-level DOI now points to the JOSS version. Does this mean that once new versions are added to the archive, the DOI in the paper will point to them instead of the version corresponding to the JOSS submission? I think this would be acceptable, but I would perhaps recommend making it clearer in the text (e.g. adding "The version corresponding to this article is XXX" on line 28).

As for the version, are `0.5.4` and `JOSS-paper` identical but for the paper itself?
