### 2021-01-28 #2823 @matthewfeickert
>  I didn't realize the top-level DOI now points to the JOSS version. Does this mean that once new versions are added to the archive, the DOI in the paper will point to them instead of the version corresponding to the JOSS submission?

Yes, the [preferred BibTeX entry for citation that we include in our README and docs](https://scikit-hep.org/pyhf/citations.html#citation) gets bumped automatically by our workflow when we make new releases to PyPI. The version number gets bumped, and the DOI stays the same. We use what Zenodo refers to as the "cite all versions" DOI for `pyhf` [`10.5281/zenodo.1169739`](https://doi.org/10.5281/zenodo.1169739) to keep people from grabbing the citation and putting it in their citation manager and citing only older versions in the future, which sadly does happen. You can see the [version release history in the Zenodo versions](https://zenodo.org/search?page=1&size=20&q=conceptrecid:%221169739%22&sort=-version&all_versions=True):

[![Zenodo_versions](https://user-images.githubusercontent.com/5142394/106194684-c4282d00-6174-11eb-8554-942f973dbe05.png)](https://zenodo.org/search?page=1&size=20&q=conceptrecid:%221169739%22&sort=-version&all_versions=True)

> I think this would be acceptable, but I would perhaps recommend making it clearer in the text (e.g. adding "The version corresponding to this article is XXX" on line 28).

Okay, the version number is already explicitly given in the citation

![draft_reference](https://user-images.githubusercontent.com/5142394/106193817-a1494900-6173-11eb-86e0-a13040e9760e.png)

but I can revise the paper to mention it in the text.

I'm a bit surprised given the metadata that exists in the left column of the first page of the paper that version number isn't listed there too. Is there a place that I can suggest revisions to the JOSS paper formattting?

> As for the version, are `0.5.4` and `JOSS-paper` identical but for the paper itself?

No, they are significantly different as we've been preparing for release `v0.6.0`.
