### 2021-01-28 #2823 @matthewfeickert
@eloisabentivegna I've incorporated the release number into the paper text now. :+1: If the preview that whedon built (above) looks okay I'll merge [`pyhf` PR 1280](https://github.com/scikit-hep/pyhf/pull/1280) to bring the changes in and then have whedon rebuild off `pyhf`'s `master` branch.
