### 2021-01-29 fdc92c9
feat: Provide support for returning correlations from minuit (#988)
* Expand minuit API to return correlations if return_correlations kwarg is True
* Provide intersphinx for iminuit docs
* Add minuit options for setting tolerance and strategy
* Add tests for returning correlations
