### 2021-01-30 #2823 @eloisabentivegna
Thanks, @matthewfeickert! The spirit of having a tagger release quoted with the paper is that it will allow readers to track down which version exactly was reviewed as part of the JOSS procedure. If `JOSS-paper` is significantly different from what @bradkav and @suchitakulkarni have reviewed, then it cannot be the release associated with the submission. Could you (and @bradkav and @suchitakulkarni) confirm that `0.5.4` is what the reviewers worked with? If so, I am happy with the new proof and ready to issue the accept command.

Regarding suggestions, you're more than welcome to open a PR on the JOSS repo (https://github.com/openjournals/joss). Thanks!
