### 2021-01-30 #2823 @matthewfeickert
> The spirit of having a tagger release quoted with the paper is that it will allow readers to track down which version exactly was reviewed as part of the JOSS procedure. If JOSS-paper is significantly different from what @bradkav and @suchitakulkarni have reviewed, then it cannot be the release associated with the submission. Could you (and @bradkav and @suchitakulkarni) confirm that 0.5.4 is what the reviewers worked with? If so, I am happy with the new proof and ready to issue the accept command.

@eloisabentivegna Thanks for the clarification. This should be fine then to use `v0.5.4`. In this case then the only thing that matters is the public API and the documentation. We only support officially released versions of `pyhf`, so as long as the public API is stable for the checks used for the review then continued development work shouldn't matter. There will be additions to the public API for `v0.6.0` that is shortly forthcoming, but nothing in the changes from the public API for `v0.5.4` would be relevant to checks in the "Functionality" section of the review (Installation, Functionality, Performance).

This is demonstrable as we have [automated nightly tests that run the test suite for the public API using the latest release from PyPI](https://github.com/scikit-hep/pyhf/actions?query=workflow%3A%22Current+Release%22) so we know for sure anytime there is a breaking change in our public API as this test breaks (c.f. [the GitHub Actions workflow here](https://github.com/scikit-hep/pyhf/blob/fdc92c90478d2ca2314fd9992bf4f1409c6a56d6/.github/workflows/release_tests.yml)).

> open a PR on the JOSS repo

Awesome! :+1: Thanks for the direction.
