### 2021-02-01 #2823 @matthewfeickert
> it's all clear now. Thanks!
>
> I will switch the version to `v0.5.4` and proceed.
>
> Thanks for your submission and interactions, and thanks @bradkav and @suchitakulkarni for finding the time for the review in this extraordinarily challenging moment.
>
> Congratulations!

Thank you so much @eloisabentivegna! You have been a fantastic editor and we very much appreciate getting to work with you on this and for how helpful you have been the whole way. :smile:

One final question: At the moment the archive is:

> Archive: [10.5281/zenodo.4472736](https://doi.org/10.5281/zenodo.4472736)

which is the `JOSS-paper` tag of `pyhf` _before_ I put in the clarifying sentence RE: the `v0.5.4` version number. Would it be better to have the archive have the exact text of the paper? If so, I can delete the current GitHub tag, and request the Zenodo team delete the current archive and then simply make a new `JOSS-paper` release on GitHub and ask `whedon` (or maybe you have to given permissions?) to update the archive URL. We're fine either way, but I in now way mind updating things and I think the Zenodo team could probably delete the current archive withing 24 hours.
