### 2021-02-01 #2823 @matthewfeickert
Thanks for updating the archive, @arfon. :+1: This all looks good to me from the `pyhf` team side now. :slightly_smiling_face:
