### 2021-02-02 #2823 @kyleniemeyer
Hello @matthewfeickert, I'm the EIC on duty this week, and doing some final checks before publishing.

I'm not sure how we missed this before, but your article is missing the required [Statement of Need](https://joss.readthedocs.io/en/latest/submitting.html#what-should-my-paper-contain) section, which illustrates the research purpose of the software. You can likely repurpose text already present in the article and just add this section header.

Please let me know when you've made that change; there is no need to redo the Zenodo archive, since that is focused on the software (JOSS itself archives the paper). Thanks!
