### 2021-02-02 #2823 @matthewfeickert
> I'm not sure how we missed this before, but your article is missing the required [Statement of Need](https://joss.readthedocs.io/en/latest/submitting.html#what-should-my-paper-contain) section, which illustrates the research purpose of the software. You can likely repurpose text already present in the article and just add this section header.

Ah that's unfortunate. It seems this is an area that should be clarified in the future as an editor and two reviewers all checked it off as being completed in the checklist, so it seems it is unclear to the review team what is required here. It is also not being enforced consistently throughout JOSS as in the last week there have been two JOSS publications missing it:

I'll work on getting this done in the next 24 hours.

> Please let me know when you've made that change; there is no need to redo the Zenodo archive, since that is focused on the software (JOSS itself archives the paper). Thanks!

@kyleniemeyer Can you please clarify more on this, as it seems that this has been a point of confusion throughout the review for everyone involved. If JOSS is responsible for archiving the paper and doesn't care if there is a dedicated Zenodo archive for the paper by the software dev team themselves, then why is there the [requirement that](https://github.com/openjournals/joss-reviews/issues/2823#issuecomment-768135982)

> The archive title should also match the submission title.

? If the goal of the Zenodo archive linked in the paper is supposed to be an archive of the _software release_ reviewed and not to capture the paper then the Zenodo archive for this paper should be the `pyhf` `v0.5.4` release: [`10.5281/zenodo.4318533`](https://zenodo.org/record/4318533) Though if this requirement about the archive title and the submission paper title matching is a hard one then we'll have to do something else as we don't want to disrupt the archive naming convention we've been using.
