### 2021-02-02 #2823 @kyleniemeyer
@matthewfeickert well, this is explicitly listed in the (short) set of items of what a paper should contain. We did change this from a suggestion to a more-formal requirement within the last year, so some papers submitted longer ago will not have it. The fact that some papers have slipped through is just due to us not being perfect. `whedon` is supposed to check for this upon submission, but currently that check fails if there are compilation issues at that stage—whedon is always getting better, but (like us) is not perfect and misses things in some edge cases.

(I'd also like to point out that this is a fairly small ask, since as I mentioned you can likely just move some of your existing text to be under this header.)

We are going to make whedon check for this more consistently, and also add the explicit requirement to the reviewer checklist to match our docs.
