### 2021-02-02 #2823 @kyleniemeyer
@matthewfeickert Regarding the version/archive question, we can certainly revisit that hard naming requirement for the archive. Ultimately our main goal is for the archive (and paper) to reflect the software reviewed.

Reviewing the history of this submission, I do see the confusion about versions... but, I don't _think_ anyone mentioned anything about the paper in the archive, just that we need to link to the version reviewed. The paper itself is not really part of the software, so changes made to that do not need to be captured in the archive. (I think things could get very circular if the paper cites the software with archived version.)

Actually, some authors even keep the paper itself in a separate branch that is never merged with the main branch, and so the software archive never contains the paper in that case.
