### 2021-02-03 #2823 @matthewfeickert
> explicitly listed in the (short) set of items of what a paper should contain.

Yes, and I know that myself, and I assume all of the reviewers in both our pre-review and review, assumed that this paper had met that given that in all circumstances the item

> **A statement of need:** Do the authors clearly state what problems the software is designed to solve and who the target audience is?

was checked off on the reviewers' checklist.

As you mention [the docs say](https://joss.readthedocs.io/en/latest/submitting.html#what-should-my-paper-contain)

> A clear _Statement of Need_ that illustrates the research purpose of the software

but the only place to mention this is meant to be an explicit **section**, and not just an idea ("is it clear from reading this why someone would need this software?") is in the following

> As this short list shows, JOSS papers are only expected to contain a limited set of metadata (see example below), a Statement of Need, Summary, Acknowledgements, and References sections

the [example paper and bibliography](https://joss.readthedocs.io/en/latest/submitting.html#example-paper-and-bibliography) doesn't make this clear either as it contains sections that are explanatory and not intended to actually appear without differentiating.

> also add the explicit requirement to the reviewer checklist to match our docs.

I think it would be quite helpful for future submissions if in the docs this could be made explicitly clear that "Statement of Need" is a **required section** of the paper.

> so some papers submitted longer ago will not have it.

This would explain why the papers we read for published examples didn't have it &mdash; thanks for clarifying.

> Regarding the version/archive question, we can certainly revisit that hard naming requirement for the archive. Ultimately our main goal is for the archive (and paper) to reflect the software reviewed.

Great &mdash; very happy to hear it. :+1: In light of that goal, @kyleniemeyer can I request that you let us keep our Zenodo naming convention and have `whedon` change the archive to be the one for the `v0.5.4` `pyhf` release: [`10.5281/zenodo.4318533`](https://zenodo.org/record/4318533)? As that's the release that was reviewed.

> I'd also like to point out that this is a fairly small ask, since as I mentioned you can likely just move some of your existing text to be under this header.

This is a very reasonable thing to ask for as an editor, and we're in no way rejecting the idea &mdash; rather we're happy for the clarification. We don't want to spend more time making continued revisions though until it is clear what the requirements for publication are and what Zenodo archive version we can use for the paper.

The Statement of Need bit is clear now so I'll tag both you and @eloisabentivegna for review of the draft of it once I have time to revise.
