### 2021-02-03 #2823 @kyleniemeyer
> I think it would be quite helpful for future submissions if in the docs this could be made explicitly clear that "Statement of Need" is a **required section** of the paper. It would help if a different example paper that included one was used as well.

Indeed, I've already submitted a [PR](https://github.com/openjournals/joss/pull/862) to correct this, and @danielskatz suggested new/improved whedon commands to check for this in https://github.com/openjournals/joss/issues/861

> Great — very happy to hear it. 👍 In light of that goal, @kyleniemeyer can I request that you let us keep our Zenodo naming convention and have `whedon` change the archive to be the one for the `v0.5.4` `pyhf` release: [`10.5281/zenodo.4318533`](https://zenodo.org/record/4318533)? As that's the release that was reviewed.

Yes, this is fine. I think we all are in favor of supporting projects that have a solid versioning and archiving process in place.
