### 2021-02-03 #2823 @bradkav
> I think it would be quite helpful for future submissions if in the docs this could be made explicitly clear that "Statement of Need" is a required section of the paper. It would help if a different example paper that included one was used as well.

I'd like to briefly jump in (as one of the reviewers) and confirm that it was not at all clear to me that the "Statement of Need" should be an explicit section (rather than just an idea to be covered in the paper). The paper did clearly explain the research purpose of the software, so I was happy to accept it.

In any case, I'm glad to see that @kyleniemeyer has submitted a [PR](https://github.com/openjournals/joss/pull/862) to clarify this.
