### 2021-02-04 8d4491f
docs: Make 'Statement of Need' explicit section in JOSS paper (#1290)
* Add 'Statement of Need' section heading to JOSS paper
   - c.f. https://github.com/openjournals/joss-reviews/issues/2823
* Move information on version and Zenodo archive back to 'Summary' section
