### 2021-02-04 #2823 @matthewfeickert
@eloisabentivegna @kyleniemeyer Can you please review the latest draft whedon has built?

Thanks also @kyleniemeyer for https://github.com/openjournals/joss/pull/862 &mdash; that's great to see! :+1:
