### 2021-02-04 #2823 @kyleniemeyer
Many thanks to @suchitakulkarni and @bradkav for reviewing this submission, and @eloisabentivegna for editing it.

(The DOI doesn't seem to be resolving yet, but the paper appears correctly at https://joss.theoj.org/papers/10.21105/joss.02823)
