### 2021-02-04 #2823 @matthewfeickert
:rocket:

> Many thanks to @suchitakulkarni and @bradkav for reviewing this submission, and @eloisabentivegna for editing it.

Many thanks to everyone indeed, and to you as well @kyleniemeyer! The `pyhf` team is deeply appreciative of your time, effort, and support to open source software. Thank you!

> (The DOI doesn't seem to be resolving yet, but the paper appears correctly at https://joss.theoj.org/papers/10.21105/joss.02823)

Sounds good. :+1:
